% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{adjust.forcing}
\alias{adjust.forcing}
\title{Adjust Forcing Parameters}
\usage{
adjust.forcing(
  Rsim.scenario,
  parameter,
  group,
  sim.year = 1,
  sim.month = 0,
  bymonth = F,
  value
)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}

\item{parameter}{Parameters to be modified (\code{ForcedEffort},\code{ForcedCatch},\code{ForcedFRate})}

\item{group}{Name of the group whose parameter is being changed. Valid values are found in the `Group` field of the object
created from running \code{rpath()}}

\item{sim.year}{Year of the simulation that should be modified.  Can be a range of years.}

\item{sim.month}{Month of the year that should be modified.  If set to 0, all months of
the year are modified.}

\item{bymonth}{Boolean value that denotes whether to use sim.year/sim.month combo
or just sim.month as a sequential vector starting at 1.}

\item{value}{New value for the parameter.}
}
\value{
Returns an Rsim.scenario object with the new parameter.
}
\description{
Modifies the various forcing parameters of the rsim scenario object.
}
\examples{
# Read in Rpath parameter file and generate balanced model
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)
# Adjust the ForcedPrey parameter for cod in year 1 for all months. Change the value to 10
Rsim.scenario.adjusted <- adjust.forcing(Rsim.scenario, parameter="ForcedPrey",group = "cod", sim.year = 1, sim.month=0,value=10)
head(Rsim.scenario.adjusted$forcing$ForcedPrey)



}
\seealso{
Other Adjust functions:
\code{\link[=adjust.fishing]{adjust.fishing()}},
\code{\link[=adjust.scenario]{adjust.scenario()}},
\code{\link[=set.rsim.scene]{set.rsim.scene()}}
}
\concept{Adjust functions}
