% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provided_drivers.R
\docType{data}
\name{asFuncs}
\alias{asFuncs}
\title{Attributespace resolution functions
A list of attribute resolution fuctions by attributespace
Provided attributespaces are 'a' for object attributes}
\format{
An object of class \code{list} of length 1.
}
\usage{
asFuncs
}
\description{
Attributespace resolution functions
A list of attribute resolution fuctions by attributespace
Provided attributespaces are 'a' for object attributes
}
\keyword{datasets}
