% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/param.R
\name{check.rpath.param}
\alias{check.rpath.param}
\title{Check Rpath parameter files}
\usage{
check.rpath.param(filename, parameter = "model")
}
\arguments{
\item{filename}{Name of the parameter file.  Must be a .csv.}

\item{parameter}{The type of parameter file you are checking.  Choices include "model",
 "diet", "juvenile", and "pedigree".}
}
\value{
Checks Rpath parameter files for consistency.  An error message will be produced if one of
 the logical checks fails.  Checks include:
 (NOTE: This does not ensure data is correct just that it is in the right places).
}
\description{
Logical check that the parameter files are filled out correctly, i.e. data is entered where it is
expected.
}
\seealso{
Other Rpath.functions: \code{\link{create.rpath.param}};
  \code{\link{ecopath}}; \code{\link{ecosim.init}};
  \code{\link{ecosim.run}}; \code{\link{webplot}};
  \code{\link{write.Rpath.sim}}; \code{\link{write.Rpath}}
}

