% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{check.rpath.params}
\alias{check.rpath.params}
\title{Check Rpath parameter files}
\usage{
check.rpath.params(Rpath.params)
}
\arguments{
\item{Rpath.params}{R object containing the parameters needed to create a
\code{Rpath} model.  This is generated either by the \code{\link{create.rpath.params}()}
 or \code{\link{read.rpath.params}()} functions.}
}
\value{
Checks Rpath parameter files for consistency.  An error message will be produced if one of
 the logical checks fails. Checks include: verification that all types are represented (e.g. consumer, producer, detrital, and fleet);
 check if input parameters were entered correctly; check if Diet columns sum to 1.
 (NOTE: This does not ensure data is correct just that it is in the right places).
}
\description{
Logical check that the parameter files are filled out correctly, i.e. data is entered where it is
expected.
}
\seealso{
Other Rpath functions:
\code{\link[=MTI]{MTI()}},
\code{\link[=create.rpath.params]{create.rpath.params()}},
\code{\link[=frate.table]{frate.table()}},
\code{\link[=read.rpath.params]{read.rpath.params()}},
\code{\link[=rpath]{rpath()}},
\code{\link[=rpath.consumers]{rpath.consumers()}},
\code{\link[=rpath.detrital]{rpath.detrital()}},
\code{\link[=rpath.gears]{rpath.gears()}},
\code{\link[=rpath.groups]{rpath.groups()}},
\code{\link[=rpath.living]{rpath.living()}},
\code{\link[=rpath.mixotrophs]{rpath.mixotrophs()}},
\code{\link[=rpath.producers]{rpath.producers()}},
\code{\link[=rpath.stanzas]{rpath.stanzas()}},
\code{\link[=stanzaplot]{stanzaplot()}},
\code{\link[=webplot]{webplot()}},
\code{\link[=write.Rpath]{write.Rpath()}},
\code{\link[=write.rpath.params]{write.rpath.params()}}
}
\concept{Rpath functions}
