% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecopath.R
\name{ecopath}
\alias{ecopath}
\title{Ecopath modual of Rpath}
\usage{
ecopath(modfile, dietfile, eco.name = NA)
}
\arguments{
\item{modfile}{Comma deliminated model parameter file.}

\item{dietfile}{Comma deliminated diet matrix file.}

\item{eco.name}{Optional name of the ecosystem which becomes an attribute of
   rpath object.}
}
\value{
Returns an Rpath object that can be supplied to the ecosim.init function.
}
\description{
Performs initial mass balance using a model parameter file and diet
matrix file.
}
\seealso{
Other Rpath.functions: \code{\link{Rsim.params}};
  \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}};
<<<<<<< HEAD
  \code{\link{ecosim.init}}; \code{\link{ecosim.pack}};
  \code{\link{ecosim.plot}}; \code{\link{ecosim.rates}};
  \code{\link{ecosim.run}}; \code{\link{ecosim.stanzas}};
  \code{\link{frate.adjust}}; \code{\link{frate.table}};
  \code{\link{webplot}}; \code{\link{write.Rpath.sim}};
  \code{\link{write.Rpath}}
=======
  \code{\link{ecosim.plot}}; \code{\link{webplot}};
  \code{\link{write.Rpath.sim}}; \code{\link{write.Rpath}}
>>>>>>> 2fe677a10d7b9e1846b962d00a99b1f73155fc7e
}

