% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ecosense.GOA}
\alias{Ecosense.GOA}
\title{Gulf of Alaska (west/central) 1990s Ecopath model (49 biological groups and 1 fleet).}
\format{
An Rpath unbalanced model object that contains:
\describe{
  \item{model}{A data.table containing Ecopath unbalanced ecosystem parameters (base parameters and fisheries).}
  \item{diet}{A data.table containing the Ecopath model's diet matrix.}
  \item{stanzas}{Parameters for multistanza groups.}
  \item{pedigree}{A data.table containing the data quality (pedigree) for the Ecopath model.}   
  ...
}
}
\source{
Whitehouse and Aydin 2020.  Assessing the sensitivity of three Alaska marine food webs to perturbations: an example of Ecosim simulations using Rpath. https://doi.org/10.1016/j.ecolmodel.2020.109074
}
\usage{
Ecosense.GOA
}
\description{
Rpath inputs (unbalanaced model).
}
\keyword{datasets}
