% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecosim.R
\name{ecosim.init.old}
\alias{ecosim.init.old}
\title{Initial set up for Ecosim modual of Rpath}
\usage{
ecosim.init.old(Rpath, juvfile, YEARS = 100)
}
\arguments{
\item{Rpath}{Rpath object containing a balanced model.}

\item{juvfile}{Comma deliminated file with multi-stanza parameters.}

\item{YEARS}{Integer value to set maximum number of years.}
}
\value{
Returns an Rpath.sim object that can be supplied to the ecosim.run function.
}
\description{
Performs initial set up for Ecosim by converting ecopath values to rates,
initializing stanzas, and packing everything together to run.
}
\seealso{
Other Rpath.functions: \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}}; \code{\link{ecopath}};
  \code{\link{ecosim.params}}; \code{\link{ecosim.plot}};
  \code{\link{ecosim.run.old}}; \code{\link{webplot}};
  \code{\link{write.Rpath.sim}}; \code{\link{write.Rpath}}
}

