% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecosim.R
\name{ecosim.pack}
\alias{ecosim.pack}
\title{Adds final variables to be used in ecosim.run}
\usage{
ecosim.pack(simpar, years, burn_years = -1, coupled = 1, crash_yr = -1)
}
\arguments{
\item{simpar}{List of ecosim parameters created by ecosim.rates and modified by ecosim.stanzas}

\item{years}{Integer value to set maximum number of years for the simulation.}

\item{burn_years}{}

\item{coupled}{}

\item{crash_yr}{}

\item{max_months}{Maximum number of months that a stanza can survive.}

\item{min_rec_factor}{Used to set minimum logistic maturity to 0.001.  Default value is ln(1 / 0.001 - 1).}
}
\value{
Returns an Rpath.sim object that can be supplied to the ecosim.run function.
}
\description{
Final step of ecosim.init that adds output matrices to the Rpath.sim object
that will be used during ecosim.run.
}
\seealso{
Other Rpath.functions: \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}}; \code{\link{ecopath}};
  \code{\link{ecosim.init}}; \code{\link{ecosim.plot}};
  \code{\link{ecosim.rates}}; \code{\link{ecosim.run}};
  \code{\link{ecosim.stanzas}}; \code{\link{frate.adjust}};
  \code{\link{frate.table}}; \code{\link{webplot}};
  \code{\link{write.Rpath.sim}}; \code{\link{write.Rpath}}
}

