% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecosim.R
\name{ecosim.run}
\alias{ecosim.run}
\title{Ecosim modual of Rpath}
\usage{
ecosim.run(simpar, BYY = 0, EYY = 0, init_run = 0)
}
\arguments{
\item{simpar}{Rpath.sim object containing ecosim parameters (Generated by ecosim.init).}

\item{BYY,}{EYY Integer values for the beginning/end year.}

\item{init_run}{Optional flag to declare if this is an inital run of the ecosim model.}
}
\value{
Returns an Rpath.sim object.
}
\description{
Runs the ecosim modual.
}
\seealso{
Other Rpath.functions: \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}}; \code{\link{ecopath}};
  \code{\link{ecosim.init}}; \code{\link{ecosim.params}};
  \code{\link{ecosim.plot}}; \code{\link{webplot}};
  \code{\link{write.Rpath.sim}}; \code{\link{write.Rpath}}
}

