% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecosim.R
\name{ecosim.stanzas}
\alias{ecosim.stanzas}
\title{Initialization of stanzas for ecosim}
\usage{
ecosim.stanzas(simpar, juvfile, steps_yr = 12, max_months = 400,
  min_rec_factor = 6.906754779)
}
\arguments{
\item{simpar}{List of ecosim parameters created by ecosim.rates}

\item{juvfile}{Comma deliminated file with multi-stanza parameters.}

\item{steps_yr}{Number of time steps per year.}

\item{max_months}{Maximum number of months that a stanza can survive.}

\item{min_rec_factor}{Used to set minimum logistic maturity to 0.001.  Default value is ln(1 / 0.001 - 1).}
}
\value{
Returns an Rpath.sim object that can be supplied to the ecosim.run function.
}
\description{
Sets up the initial conditions for groups with multiple stanzas.
}
\seealso{
Other Rpath.functions: \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}}; \code{\link{ecopath}};
  \code{\link{ecosim.init}}; \code{\link{ecosim.pack}};
  \code{\link{ecosim.plot}}; \code{\link{ecosim.rates}};
  \code{\link{ecosim.run}}; \code{\link{webplot}};
  \code{\link{write.Rpath.sim}}; \code{\link{write.Rpath}}
}

