% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{frate.adjust}
\alias{frate.adjust}
\title{Adjust Fishing Mortality}
\usage{
frate.adjust(Rpath.sim, group, gear, frate, multiplier = T, target = T,
  discard = T)
}
\arguments{
\item{group}{Name or number of the species group whose fishing mortality rate
is being changed.}

\item{gear}{Name or number of the gear to which the change applies.}

\item{frate}{New fishing mortality rate.  Can be an actual value or a multiplier.}

\item{multiplier}{Logical value indicating whether or not the frate is a multiplier.}

\item{target}{Logical value indicating whether the change effects landings mortality.}

\item{discard}{Logical value indicating whether the change effects discard mortality.}

\item{Rsim.scenario}{Object generated by ecosim.init.}
}
\value{
Returns an Rpath.sim object with the new F rate.
}
\description{
Modifies the fishing mortality value for a species by a particular gear.
}
\seealso{
Other Rpath.functions: \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}};
  \code{\link{frate.table}}; \code{\link{rpath.stanzas}};
  \code{\link{rpath}}; \code{\link{rsim.params}};
  \code{\link{rsim.plot}}; \code{\link{stanzaplot}};
  \code{\link{webplot}}; \code{\link{write.Rpath.sim}};
  \code{\link{write.Rpath}}
}

