% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{frate.table}
\alias{frate.table}
\title{Fishing Mortality Table}
\usage{
frate.table(Rsim.scenario)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}
}
\value{
Returns a data table of F values for each species/gear combination.
\item{Group}{Group/node}
\item{Fishery/gearType}{Gear type for the fishery} ...
\item{Total}{Sum over all gear types}
}
\description{
Creates a table of fishing mortalities by species group and gear for an
\code{rsim.scenario()} object.
}
\examples{
# Read in Rpath parameter file and generate balanced model
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)
rates <- frate.table(Rsim.scenario)
# display the head of the data frame
head(rates)

}
\seealso{
Other Rpath functions:
\code{\link[=MTI]{MTI()}},
\code{\link[=check.rpath.params]{check.rpath.params()}},
\code{\link[=create.rpath.params]{create.rpath.params()}},
\code{\link[=read.rpath.params]{read.rpath.params()}},
\code{\link[=rpath]{rpath()}},
\code{\link[=rpath.consumers]{rpath.consumers()}},
\code{\link[=rpath.detrital]{rpath.detrital()}},
\code{\link[=rpath.gears]{rpath.gears()}},
\code{\link[=rpath.groups]{rpath.groups()}},
\code{\link[=rpath.living]{rpath.living()}},
\code{\link[=rpath.mixotrophs]{rpath.mixotrophs()}},
\code{\link[=rpath.producers]{rpath.producers()}},
\code{\link[=rpath.stanzas]{rpath.stanzas()}},
\code{\link[=stanzaplot]{stanzaplot()}},
\code{\link[=webplot]{webplot()}},
\code{\link[=write.Rpath]{write.Rpath()}},
\code{\link[=write.rpath.params]{write.rpath.params()}}
}
\concept{Rpath functions}
