% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{get.rsim.fishing}
\alias{get.rsim.fishing}
\title{Retrieve fishing forcing parameters from an Rsim scenario}
\usage{
get.rsim.fishing(Rsim.scenario)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}
}
\value{
Returns a `fishing` object.
}
\description{
Helper function that will retrieve the fishing forcing parameters that were used in an
Rsim scenario
}
\examples{
# Read in Rpath parameter file and generate balanced model
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)
params <- get.rsim.fishing(Rsim.scenario)
names(params)



}
\seealso{
Other Get functions:
\code{\link[=get.rsim.forcing]{get.rsim.forcing()}},
\code{\link[=get.rsim.params]{get.rsim.params()}},
\code{\link[=get.rsim.stanzas]{get.rsim.stanzas()}},
\code{\link[=get.rsim.start_state]{get.rsim.start_state()}}
}
\concept{Get functions}
