% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpath_support.R
\name{grouptype}
\alias{grouptype}
\title{Set of functions for returning functional group names (character vector)
using the type input column
Internal supporting function for group names to check type}
\usage{
grouptype(RpathObj)
}
\arguments{
\item{RpathObj}{An Rpath object}
}
\value{
Returns a list of the type and group associated with the passed Rpath object
}
\description{
Set of functions for returning functional group names (character vector)
using the type input column
Internal supporting function for group names to check type
}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{adjust.forcing}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{extract.node}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.stanzas}()},
\code{\link{rpath}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.Rsim}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
