% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provided_drivers.R
\name{list_termination}
\alias{list_termination}
\title{List termination condition
Termination condition which traverses to the first non-list (e.g. a vector) element}
\usage{
list_termination(x)
}
\description{
List termination condition
Termination condition which traverses to the first non-list (e.g. a vector) element
}
