% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{as.list.rpath_matchList}
\alias{as.list.rpath_matchList}
\title{conversion Methods for rpath_matchList

These are an internal detail which should not be relevant to end users}
\usage{
\method{as.list}{rpath_matchList}(from)
}
\arguments{
\item{from}{Object being converted.}
}
\value{
a list, or an rpath_matchList object, depending on direction of the coercion.
}
\description{
conversion Methods for rpath_matchList

These are an internal detail which should not be relevant to end users
}
