% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{makeParsers}
\alias{makeParsers}
\title{^([^\\[!=&\\(\\|]+|\\([^\\)]+\\))(==|!=|\\|\\||&&)([^\\[!=&\\|\\(]+|\\([^\\)]+\\))$' = function(match, index, state)}
\usage{
makeParsers(state, ...)
}
\description{
^([^\\[!=&\\(\\|]+|\\([^\\)]+\\))(==|!=|\\|\\||&&)([^\\[!=&\\|\\(]+|\\([^\\)]+\\))$' = function(match, index, state)
}
