% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_functions.R
\name{MTI}
\alias{MTI}
\title{Quantify mixed trophic impacts}
\usage{
MTI(Rpath, Rpath.params, increase = T)
}
\arguments{
\item{Rpath}{Rpath R object containing a static \code{Rpath} model.}

\item{Rpath.params}{R object containing the \code{Rpath} model parameters.  This is generated
either by the \code{\link{create.rpath.params}()} or \code{\link{read.rpath.params}()} functions.}

\item{increase}{Logical value indicating whether a marginal increase is applied.}
}
\value{
Returns a matrix of mixed trophic impacts.
}
\description{
Uses a static \code{Rpath} model and creates a matrix of mixed trophic impacts.
}
\seealso{
Other Rpath functions:
\code{\link[=check.rpath.params]{check.rpath.params()}},
\code{\link[=create.rpath.params]{create.rpath.params()}},
\code{\link[=frate.table]{frate.table()}},
\code{\link[=read.rpath.params]{read.rpath.params()}},
\code{\link[=rpath]{rpath()}},
\code{\link[=rpath.consumers]{rpath.consumers()}},
\code{\link[=rpath.detrital]{rpath.detrital()}},
\code{\link[=rpath.gears]{rpath.gears()}},
\code{\link[=rpath.groups]{rpath.groups()}},
\code{\link[=rpath.living]{rpath.living()}},
\code{\link[=rpath.mixotrophs]{rpath.mixotrophs()}},
\code{\link[=rpath.producers]{rpath.producers()}},
\code{\link[=rpath.stanzas]{rpath.stanzas()}},
\code{\link[=stanzaplot]{stanzaplot()}},
\code{\link[=webplot]{webplot()}},
\code{\link[=write.Rpath]{write.Rpath()}},
\code{\link[=write.rpath.params]{write.rpath.params()}}
}
\concept{Rpath functions}
