% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provided_drivers.R
\docType{data}
\name{nsFuncs}
\alias{nsFuncs}
\title{Namespace resolution functions
A list of node-name resolution functions by namespace
Provided namespaces are 'cl' for element class and nm for element name (elements with  no name are given name "", which will never match a non-wildcard (*) rpath directive.}
\format{
An object of class \code{list} of length 2.
}
\usage{
nsFuncs
}
\description{
Namespace resolution functions
A list of node-name resolution functions by namespace
Provided namespaces are 'cl' for element class and nm for element name (elements with  no name are given name "", which will never match a non-wildcard (*) rpath directive.
}
\keyword{datasets}
