% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecofitting.R
\name{read.fitting.biomass}
\alias{read.fitting.biomass}
\title{Read and Process Biomass Fitting Data}
\usage{
read.fitting.biomass(Rsim.scenario, biomass_data)
}
\arguments{
\item{Rsim.scenario}{An `Rsim` scenario object containing simulation parameters and setup.}

\item{biomass_data}{A `data.frame` containing biomass observation data.}
}
\value{
Modifies and returns a new Rsim scenario object, attaching the cleaned data frame to 
  `Rsim.scenario$fitting$Biomass` with additional calculated columns (`obs`, `sd`, `initial_q`, `wt`).
}
\description{
Reads, cleans, and formats biomass observation data to prepare it for model 
fitting within an `Rsim` scenario object.
}
\details{
Observation data must be provided in the `biomass_data` data frame.  `biomass_data` 
must contain the following columns: `Group`, `Year`, `Value`, `Scale`, `Source`, 
one of either `Stdev` or `CV` (or both), and optionally `Type`. If both `Stdev`
and `CV` are provided, `Stdev` is used for each data point unless it is missing (NA)
in which case `CV` is used.
}
