% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{read.rpath.params}
\alias{read.rpath.params}
\title{Read Rpath parameters from .csv files}
\usage{
read.rpath.params(
  modfile,
  dietfile,
  pedfile = NA,
  stanzagroupfile = NA,
  stanzafile = NA
)
}
\arguments{
\item{modfile}{file location of the flat file containing the \emph{model} parameters.}

\item{dietfile}{file location of the flat file containing the \emph{diet} parameters.}

\item{pedfile}{file location of the flat file containing the \emph{pedgigree} parameters.}

\item{stanzagroupfile}{file location of the flat file containing the \emph{group} parameters
for multistanza groups. If not specified a blank stanza list will be created.}

\item{stanzafile}{file location of the flat file containing the individual stanza
parameters for multistanza groups. If not specified a blank stanza list will
be created.}
}
\value{
Outputs an \code{Rpath.param} object that can be used for Rpath and subsequently
 Rsim.  (NOTE: This does function does not ensure data is correct or in the
 correct locations.Run \code{check.rpath.params()} to ensure the appropriate columns are
 present).
}
\description{
Creates an \code{Rpath.param} object from a series of .csv files.
}
\seealso{
Other Rpath functions:
\code{\link[=MTI]{MTI()}},
\code{\link[=check.rpath.params]{check.rpath.params()}},
\code{\link[=create.rpath.params]{create.rpath.params()}},
\code{\link[=frate.table]{frate.table()}},
\code{\link[=rpath]{rpath()}},
\code{\link[=rpath.consumers]{rpath.consumers()}},
\code{\link[=rpath.detrital]{rpath.detrital()}},
\code{\link[=rpath.gears]{rpath.gears()}},
\code{\link[=rpath.groups]{rpath.groups()}},
\code{\link[=rpath.living]{rpath.living()}},
\code{\link[=rpath.mixotrophs]{rpath.mixotrophs()}},
\code{\link[=rpath.producers]{rpath.producers()}},
\code{\link[=rpath.stanzas]{rpath.stanzas()}},
\code{\link[=stanzaplot]{stanzaplot()}},
\code{\link[=webplot]{webplot()}},
\code{\link[=write.Rpath]{write.Rpath()}},
\code{\link[=write.rpath.params]{write.rpath.params()}}
}
\concept{Rpath functions}
