% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpath_support.R
\name{rpath.gears}
\alias{rpath.gears}
\title{List of Rpath gears groups}
\usage{
rpath.gears(Rpath)
}
\arguments{
\item{Rpath}{Balanced Rpath model generated by \code{rpath()}}
}
\value{
Returns a string vector containing the names of Rpath functional groups
that are gear types
}
\description{
List of gears groups from Rpath object with species type = 3
}
\examples{
# Build the balanced Rpath model and parameter file by calling `rpath`
Rpath <- rpath(Ecosense.EBS)
# Run the groups command
gearGroups = rpath.gears(Rpath)
# Print out the first few gear type group names
head(gearGroups)

}
\seealso{
Other Rpath functions:
\code{\link[=MTI]{MTI()}},
\code{\link[=check.rpath.params]{check.rpath.params()}},
\code{\link[=create.rpath.params]{create.rpath.params()}},
\code{\link[=frate.table]{frate.table()}},
\code{\link[=read.rpath.params]{read.rpath.params()}},
\code{\link[=rpath]{rpath()}},
\code{\link[=rpath.consumers]{rpath.consumers()}},
\code{\link[=rpath.detrital]{rpath.detrital()}},
\code{\link[=rpath.groups]{rpath.groups()}},
\code{\link[=rpath.living]{rpath.living()}},
\code{\link[=rpath.mixotrophs]{rpath.mixotrophs()}},
\code{\link[=rpath.producers]{rpath.producers()}},
\code{\link[=rpath.stanzas]{rpath.stanzas()}},
\code{\link[=stanzaplot]{stanzaplot()}},
\code{\link[=webplot]{webplot()}},
\code{\link[=write.Rpath]{write.Rpath()}},
\code{\link[=write.rpath.params]{write.rpath.params()}}
}
\concept{Rpath functions}
