% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpath_support.R
\name{rpath.groups}
\alias{rpath.groups}
\title{Rpath functional group names}
\usage{
rpath.groups(Rpath)
}
\arguments{
\item{Rpath}{Balanced Rpath model generated by \code{rpath()}}
}
\value{
Returns a string vector containing the names of Rpath functional groups
by category (group type).
}
\description{
Get a character vector of functional group names from an Rpath object (balanced model)  
or an Rpath.params object (unbalanced model parameters) based on the 'type' input
parameter as follows: (0: consumers, 1: producers, 2: detrital, 3: gears, 0<type<1: mixotrophs).
Living groups are consumers + producers.  Note that mixotrophs are not returned as
either consumers or producers, only separately.
}
\examples{
# Build the balanced Rpath model and parameter file by calling `rpath`
Rpath <- rpath(Ecosense.EBS)
# Run the groups command
groups = rpath.groups(Rpath)
# Print out the first few group names
head(groups)

}
\seealso{
Other Rpath functions:
\code{\link[=MTI]{MTI()}},
\code{\link[=check.rpath.params]{check.rpath.params()}},
\code{\link[=create.rpath.params]{create.rpath.params()}},
\code{\link[=frate.table]{frate.table()}},
\code{\link[=read.rpath.params]{read.rpath.params()}},
\code{\link[=rpath]{rpath()}},
\code{\link[=rpath.consumers]{rpath.consumers()}},
\code{\link[=rpath.detrital]{rpath.detrital()}},
\code{\link[=rpath.gears]{rpath.gears()}},
\code{\link[=rpath.living]{rpath.living()}},
\code{\link[=rpath.mixotrophs]{rpath.mixotrophs()}},
\code{\link[=rpath.producers]{rpath.producers()}},
\code{\link[=rpath.stanzas]{rpath.stanzas()}},
\code{\link[=stanzaplot]{stanzaplot()}},
\code{\link[=webplot]{webplot()}},
\code{\link[=write.Rpath]{write.Rpath()}},
\code{\link[=write.rpath.params]{write.rpath.params()}}
}
\concept{Rpath functions}
