% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpath_support.R
\name{rpath.producers}
\alias{rpath.producers}
\title{List of Rpath producer groups}
\usage{
rpath.producers(Rpath)
}
\arguments{
\item{Rpath}{Balanced Rpath model generated by \code{rpath()}}
}
\value{
Returns a string vector containing the names of Rpath functional groups
that are producers
}
\description{
List of producer groups from Rpath object with species type = 1
}
\examples{
# Build the balanced Rpath model and parameter file by calling `rpath`
Rpath <- rpath(Ecosense.EBS)
# Run the groups command
producerGroups = rpath.producers(Rpath)
# Print out the first few producer group names
head(producerGroups)

}
\seealso{
Other Rpath functions:
\code{\link[=MTI]{MTI()}},
\code{\link[=check.rpath.params]{check.rpath.params()}},
\code{\link[=create.rpath.params]{create.rpath.params()}},
\code{\link[=frate.table]{frate.table()}},
\code{\link[=read.rpath.params]{read.rpath.params()}},
\code{\link[=rpath]{rpath()}},
\code{\link[=rpath.consumers]{rpath.consumers()}},
\code{\link[=rpath.detrital]{rpath.detrital()}},
\code{\link[=rpath.gears]{rpath.gears()}},
\code{\link[=rpath.groups]{rpath.groups()}},
\code{\link[=rpath.living]{rpath.living()}},
\code{\link[=rpath.mixotrophs]{rpath.mixotrophs()}},
\code{\link[=rpath.stanzas]{rpath.stanzas()}},
\code{\link[=stanzaplot]{stanzaplot()}},
\code{\link[=webplot]{webplot()}},
\code{\link[=write.Rpath]{write.Rpath()}},
\code{\link[=write.rpath.params]{write.rpath.params()}}
}
\concept{Rpath functions}
