% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopath.R
\name{rpath}
\alias{rpath}
\title{Ecopath module of Rpath}
\usage{
rpath(Rpath.params, eco.name = NA, eco.area = 1)
}
\arguments{
\item{Rpath.params}{R object containing the parameters needed to create a
\code{Rpath} model.  This is generated either by the \code{\link{create.rpath.params}()}
 or \code{\link{read.rpath.params}()} functions.}

\item{eco.name}{Optional name of the ecosystem which becomes an attribute of the
\code{Rpath} model.}

\item{eco.area}{Optional area of the ecosystem which becomes an attribute of the
\code{Rpath} model.}
}
\value{
Returns a static \code{Rpath} model that can be supplied to the
\code{rsim.scenario} function.
}
\description{
Performs initial mass balance using a \code{Rpath.params()} file
}
\seealso{
Other Rpath functions:
\code{\link[=MTI]{MTI()}},
\code{\link[=check.rpath.params]{check.rpath.params()}},
\code{\link[=create.rpath.params]{create.rpath.params()}},
\code{\link[=frate.table]{frate.table()}},
\code{\link[=read.rpath.params]{read.rpath.params()}},
\code{\link[=rpath.consumers]{rpath.consumers()}},
\code{\link[=rpath.detrital]{rpath.detrital()}},
\code{\link[=rpath.gears]{rpath.gears()}},
\code{\link[=rpath.groups]{rpath.groups()}},
\code{\link[=rpath.living]{rpath.living()}},
\code{\link[=rpath.mixotrophs]{rpath.mixotrophs()}},
\code{\link[=rpath.producers]{rpath.producers()}},
\code{\link[=rpath.stanzas]{rpath.stanzas()}},
\code{\link[=stanzaplot]{stanzaplot()}},
\code{\link[=webplot]{webplot()}},
\code{\link[=write.Rpath]{write.Rpath()}},
\code{\link[=write.rpath.params]{write.rpath.params()}}
}
\concept{Rpath functions}
