% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.diet}
\alias{rsim.diet}
\title{Output consumption by a group}
\usage{
rsim.diet(Rsim.output, group)
}
\arguments{
\item{Rsim.output}{R object containing the output from \code{rsim.run}.}

\item{group}{Group from the \code{Rpath} model that is of interest}
}
\description{
Creates a matrix of consumption of prey by a particular predator.
}
