% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecofitting.R
\name{rsim.fit.list.catch.series}
\alias{rsim.fit.list.catch.series}
\title{List Catch Timeseries Groups}
\usage{
rsim.fit.list.catch.series(Rsim.scenario)
}
\arguments{
\item{Rsim.scenario}{An `Rsim` scenario list object containing 
fitting data setup.}
}
\value{
A list containing a single element, `groups`, which holds a vector 
  of the unique group names present in the catch data.
}
\description{
Extracts the unique species or functional groups that have observation data 
within the catch fitting matrix of an `Rsim` simulation scenario.
}
\details{
This is a helper function designed to quickly identify which groups have 
empirical catch data loaded into the scenario. It reads the `Group` column 
from the `Rsim.scenario$fitting$Catch` data frame and returns the unique values.
}
