% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim_plot.R
\name{rsim.plot}
\alias{rsim.plot}
\title{Plot routine for Rsim runs}
\usage{
rsim.plot(Rsim.output, spname = "all", indplot = F)
}
\arguments{
\item{Rsim.output}{R object containing the output from \code{rsim.run()}.}

\item{spname}{Vector of species names to be displayed.  The default "all" will
display all group names}

\item{indplot}{Logical value of whether to plot a single group or multiple groups}
}
\value{
Creates a figure of relative biomass
}
\description{
Plots the relative biomass of each group from an \code{Rsim.output} object.
}
\section{Details}{


For a detailed example of how to use this plot command please see \code{vignette("RunRsim")}
}

\seealso{
Other Rsim functions:
\code{\link[=extract.node]{extract.node()}},
\code{\link[=rsim.params]{rsim.params()}},
\code{\link[=rsim.run]{rsim.run()}},
\code{\link[=rsim.scenario]{rsim.scenario()}},
\code{\link[=rsim.step]{rsim.step()}},
\code{\link[=write.Rsim]{write.Rsim()}}
}
\concept{Rsim functions}
