% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecofitting.R
\name{rsim.predprey.table}
\alias{rsim.predprey.table}
\title{Extract Predator-Prey Interaction Table}
\usage{
rsim.predprey.table(Rsim.scenario, pred = NULL, prey = NULL)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}

\item{pred}{An optional character string or vector of predator names. If not
provided, the function will include all model predators in its output.}

\item{prey}{An optional character string or vector of prey names. If not
provided, the function will include all model prey in its output.}
}
\value{
A `data.frame` containing the columns: 
\itemize{
  \item{\code{from}, prey name}, 
  \item{\code{to} predator name},
  \item{\code{QQ} base consumption rate of the predator/prey link},
  \item{\code{VV} vulnerability parameter of the predator/prey link},
  \item{\code{DD} handling time parameter of the predator/prey link}
}
}
\description{
Generates a data frame of predator-prey linkages and their 
associated foraging parameters from an `Rpath` simulation scenario.
}
\details{
This function outputs a data frame of predator/prey link parameters contained 
in an Rsim scenario for a given set of predators and prey.
}
