% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.run}
\alias{rsim.run}
\title{Run a dynamic simulation}
\usage{
rsim.run(Rsim.scenario, method = "RK4", years = 1:100, spname = NULL)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}

\item{method}{Numerical integration method. Either \emph{'AB'} for Adams-Bashforth or 
\emph{'RK4'} for 4th order Runge-Kutta.}

\item{years}{A vector of each year of the simulation.}
}
\value{
S3 object class \code{Rsim.output}, a list of 16 objects:
\itemize{
 \item{\code{out_Biomass}, numeric matrix of monthly (rows) output biomass values by biomass group (columns) for model run}
 \item{\code{out_Catch}, numeric matrix of monthly (rows) output catch values by biomass group (columns) for model run}
 \item{\code{out_Gear_Catch}, numeric matrix of monthly (rows) output catch by species/gear/disposition (columns) for model run}
 \item{\code{annual_Catch}, numeric matrix of annual (rows) output catch values by biomass group (columns) for model run}
 \item{\code{annual_Biomass}, numeric matrix of annual (rows) output biomass values by biomass group (columns) for model run}
 \item{\code{annual_QB}, numeric matrix of annual (rows) output consumption:biomass values by biomass group (columns) for model run}
 \item{\code{annual_Qlink}, numeric matrix of annual (rows) output consumption values by predator-prey pair (columns) for model run}
 \item{\code{end_state}, list of 9 final value vectors for biomass groups, same components as those produced by \code{\link{rsim.state}()}}
 \item{\code{crash_year}, integer flag; -1 no crash or year of model crash}
 \item{\code{pred}, named character vector of predators in each predator prey pair}
 \item{\code{prey}, named character vector of prey in each predator prey pair}
 \item{\code{Gear_Catch_sp}, named character vector of species caught in gears for each disposition}
 \item{\code{Gear_Catch_gear}, named character vector of gear types for each species and disposition}
 \item{\code{Gear_Catch_disp}, character vector of catch disposition (Landing or Discard) for each species and gear}
 \item{\code{start_state}, list of 9 initial value vectors for biomass groups, see output of \code{\link{rsim.state}()}}
 \item{\code{params}, list of 4; 3 dimensioning parameters and 1 biomass group and fleet name vector}
}
}
\description{
Carries out the numerical integration of the Rsim alogrithms.
}
\examples{
# Read in Rpath parameter file and generate model object
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)
# Run the Rsim simulation
Rsim.output <- rsim.run(Rsim.scenario, method = "RK4", years = 1:50)
  
}
\seealso{
Other Rsim functions:
\code{\link[=extract.node]{extract.node()}},
\code{\link[=rsim.params]{rsim.params()}},
\code{\link[=rsim.plot]{rsim.plot()}},
\code{\link[=rsim.scenario]{rsim.scenario()}},
\code{\link[=rsim.step]{rsim.step()}},
\code{\link[=write.Rsim]{write.Rsim()}}
}
\concept{Rsim functions}
