% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.scenario}
\alias{rsim.scenario}
\title{Specify a dynamic simulation scenario}
\usage{
rsim.scenario(Rpath, Rpath.params, years = 1:100)
}
\arguments{
\item{Rpath}{R object containing a static \code{Rpath} model.}

\item{Rpath.params}{R object containing the parameters needed to create a
\code{Rpath} model.  This is generated either by the \code{\link{create.rpath.params}()}
 or \code{\link{read.rpath.params}()} functions.}

\item{years}{A vector of each year of the simulation.}
}
\value{
Returns an \code{Rsim.scenario} object that can be supplied to the 
\code{\link{rsim.run}()} function. \code{Rsim.scenario} is a list of 5 objects:
\itemize{
 \item{\code{params}, list of 39 dynamic parameters and vectors derived from an \code{Rpath} model, see output of \code{\link{rsim.params}()}}
 \item{\code{start_state}, list of 9 initial value vectors for biomass and size structured stanza groups, see output of \code{\link{rsim.state}()} and \code{\link{rsim.stanzas}()}}
 \item{\code{forcing}, list of 7 matrices for time series forcing of biological attributes, see output of \code{\link{rsim.forcing}()}}
 \item{\code{fishing}, list of 3 matrices for time series forcing of fishing attributes, see output of \code{\link{rsim.fishing}()}}
 \item{\code{stanzas}, list of 21 parameters and vectors for age structured groups, see output of \code{\link{rsim.stanzas}()}}
}
}
\description{
Uses a static \code{Rpath} model and creates a scenario consisting of 5 list objects:
\code{params} generated by \code{\link{rsim.params}()}, \code{start_state} generated by \code{\link{rsim.state}()}, 
\code{forcing} generated by \code{\link{rsim.forcing}()}, \code{fishing} generated by \code{\link{rsim.fishing}()}, 
and '\code{stanzas} generated by \code{\link{rsim.stanzas}()}.
}
\examples{
# Read in Rpath parameter file and generate model object
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)

}
\seealso{
Other Rsim functions:
\code{\link[=extract.node]{extract.node()}},
\code{\link[=rsim.params]{rsim.params()}},
\code{\link[=rsim.plot]{rsim.plot()}},
\code{\link[=rsim.run]{rsim.run()}},
\code{\link[=rsim.step]{rsim.step()}},
\code{\link[=write.Rsim]{write.Rsim()}}
}
\concept{Rsim functions}
