% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosense.R
\name{rsim.sense.params}
\alias{rsim.sense.params}
\title{Ecosense modual of Rpath}
\usage{
rsim.sense.params(Rpath.scenario, Rpath, Rpath.params, mscramble = 2,
  mhandle = 1000, preyswitch = 1, scrambleselfwt = 1, handleselfwt = 1,
  steps_yr = 12, steps_m = 1)
}
\arguments{
\item{Rpath.scenario}{R object that is generated by the rsim.scenario
function and can be supplied to the rsim.run function.}

\item{Rpath}{R object containing a balanced Rpath model.}

\item{Rpath.params}{R object containing the Rpath parameters.  This is generated
either by the create.rpath.params or read.rpath.params functions.}
}
\value{
Returns an Rsim.scenario object that can be supplied to the rsim.run function.
}
\description{
Ecosense modual of Rpath
}
\seealso{
Other Rpath.functions: \code{\link{adjust.fishing}},
  \code{\link{adjust.scenario}},
  \code{\link{check.rpath.params}},
  \code{\link{create.rpath.params}},
  \code{\link{extract.node}}, \code{\link{frate.table}},
  \code{\link{read.rpath.params}},
  \code{\link{rpath.stanzas}}, \code{\link{rpath}},
  \code{\link{rsim.params}}, \code{\link{rsim.plot}},
  \code{\link{rsim.scenario}}, \code{\link{stanzaplot}},
  \code{\link{webplot}}, \code{\link{write.Rpath}},
  \code{\link{write.Rsim}},
  \code{\link{write.rpath.params}}
}

