% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosense.R
\name{rsim.sense.path}
\alias{rsim.sense.path}
\title{Ecosense function for rpath}
\usage{
rsim.sense.path(Rsim.scenario, Rpath, Rpath.params, steps_yr = 12, steps_m = 1)
}
\arguments{
\item{Rsim.scenario}{The \code{rsim.scenario()} object containing the initial
rsim parameters used as the center of distributions (base model) during
random parameter generation.}

\item{Rpath}{The balanced Rpath model created by \code{rpath()}.}

\item{Rpath.params}{Rpath parameter object (unbalanced model) containing the
data pedigree generated by either \code{create.rpath.params()} or
\code{read.rpath.params()}.}

\item{steps_yr}{Specifies the number time steps per year in the
\code{rsim.scenario()}}

\item{steps_m}{Specifies the number of time steps per month in the
\code{rsim.scenario()}}
}
\value{
Returns a \code{params} list object that can be substituted for the
 \code{params} list in an \code{rsim.scenario} object.
}
\description{
Warning: \code{rsim.sense.path()} is deprecated and will be
 removed in future versions. It is included here for testing purposes only.
 Please update your code to use the preferred function \code{rsim.sense()}.

 \code{rsim.sense.path()} draws a random Rsim parameter set from
 distributions centered on the supplied Rpath parameter estimates. The width
 of the distribution for each parameter is defined by the data pedigree of
 the \code{Rpath.params} object generated by either
 \code{create.rpath.params()} or \code{read.rpath.params()}. The returned
 Rsim parameter set is functionally equivalent to a \code{params} list object
 created by \code{rsim.scenario()}.
}
\references{
Whitehouse GA, Aydin KY (2020) Assessing the sensitivity of three
 Alaska marine food webs to perturbations: an example of Ecosim simulations
 using Rpath. Ecol Model 429:16. doi:10.1016/j.ecolmodel.2020.109074
}
\seealso{
\code{vignette("ecosense", package = "Rpath")} for use of
 \code{rsim.sense()}, and see other Rpath functions:
 \code{\link{read.rpath.params}()}, \code{\link{create.rpath.params}()},
 \code{\link{check.rpath.params}()}, \code{\link{rpath}()},
 \code{\link{rsim.scenario}()}, \code{\link{rsim.run}()}
}
\concept{Rpath functions

  rsim.sense.path is depreciated and included for testing only, the preferred
  function is rsim.sense(). Whitehouse and Aydin (Submitted) Assessing the
  sensitivity of three Alaska marine food webs to perturbations: an example
  of Ecosim simulations using Rpath

  This function generates random parameters around the Ecopath baseline, not
  the scenario.}
