% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.stanzas}
\alias{rsim.stanzas}
\title{Generate Rsim stanza matrix}
\usage{
rsim.stanzas(Rpath.params, state, params)
}
\arguments{
\item{Rpath.params}{R object containing the Rpath parameters.  This is generated
either by the create.rpath.params or read.rpath.params functions.}

\item{state}{List object of state variables generated by \code{rsim.state}.}

\item{params}{Rsim parameter object generated by \code{rsim.params}.}
}
\value{
S3 object class \code{Rsim.state}, a list of 21 objects:
\itemize{
 \item{\code{SpawnBio}, numeric vector of initial age structured stanza }
 \item{\code{StanzaPred}, numeric vector}
 \item{\code{EggsStanza}, numeric vector}
 \item{\code{NageS}, numeric matrix of monthly (rows) biomass multiplier by biomass group (columns)}
 \item{\code{WageS}, numeric matrix of monthly (rows) biomass multiplier by biomass group (columns)}
 \item{\code{QageS}, numeric matrix of monthly (rows) biomass multiplier by biomass group (columns)}
}
}
\description{
Creates a matrix of stanza variables to be used by \code{rsim.run}.
}
