% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.state}
\alias{rsim.state}
\title{Generate Rsim state matrix}
\usage{
rsim.state(params)
}
\arguments{
\item{params}{Rsim parameter object generated by \code{rsim.params}.}
}
\description{
Creates a matrix of state variables used by Rsim.
}
