% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim_multistep.R
\name{rsim.step}
\alias{rsim.step}
\title{Function to add a time step to an Rsim run}
\usage{
rsim.step(Rsim.scenario, Rsim.output, method = "AB", year.end)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}

\item{Rsim.output}{R object containing the output from \code{rsim.run}.}

\item{method}{Numerical integration method. Either \emph{'AB'} for Adams-Bashforth or 
\emph{'RK4'} for 4th order Runge-Kutta.}

\item{year.end}{The final year for the step interval}
}
\value{
Returns an \code{Rsim.output} object.
}
\description{
Runs \code{rsim.run} and appends the output to a previous \code{Rsim.output}.
Currently only works with the Adams-Bashforth numerical integration method (AB).
}
\seealso{
Other Rsim functions:
\code{\link[=extract.node]{extract.node()}},
\code{\link[=rsim.params]{rsim.params()}},
\code{\link[=rsim.plot]{rsim.plot()}},
\code{\link[=rsim.run]{rsim.run()}},
\code{\link[=rsim.scenario]{rsim.scenario()}},
\code{\link[=write.Rsim]{write.Rsim()}}
}
\concept{Rsim functions}
