% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopath_plot.R
\name{stanzaplot}
\alias{stanzaplot}
\title{Plot routine for Ecopath multistanzas}
\usage{
stanzaplot(
  Rpath.params,
  StanzaGroup,
  line.cols = c("black", "green", "blue", "red")
)
}
\arguments{
\item{Rpath.params}{R object containing the parameters needed to create a
\code{Rpath} model.  This is generated either by the \code{\link{create.rpath.params}()}
 or \code{\link{read.rpath.params}()} functions.}

\item{StanzaGroup}{The Stanza group's name to be plotted.Valid values are found in the 
`stanzas` field of the Rpath parameter file.}

\item{line.cols}{A vector of four colors used to represent the population biomass,
relative number, individual weights, and stanza separation lines.}
}
\value{
Creates a figure showing the break down of biomass and number per stanza.
}
\description{
Plots the biomass composition of multistanza groups from an Rpath.stanzas object.
}
\examples{
# Choose group with multiple stanzas
params <- REco.params
stanzaplot(params, StanzaGroup = "Roundfish1")


}
\seealso{
Other Rpath functions:
\code{\link[=MTI]{MTI()}},
\code{\link[=check.rpath.params]{check.rpath.params()}},
\code{\link[=create.rpath.params]{create.rpath.params()}},
\code{\link[=frate.table]{frate.table()}},
\code{\link[=read.rpath.params]{read.rpath.params()}},
\code{\link[=rpath]{rpath()}},
\code{\link[=rpath.consumers]{rpath.consumers()}},
\code{\link[=rpath.detrital]{rpath.detrital()}},
\code{\link[=rpath.gears]{rpath.gears()}},
\code{\link[=rpath.groups]{rpath.groups()}},
\code{\link[=rpath.living]{rpath.living()}},
\code{\link[=rpath.mixotrophs]{rpath.mixotrophs()}},
\code{\link[=rpath.producers]{rpath.producers()}},
\code{\link[=rpath.stanzas]{rpath.stanzas()}},
\code{\link[=webplot]{webplot()}},
\code{\link[=write.Rpath]{write.Rpath()}},
\code{\link[=write.rpath.params]{write.rpath.params()}}
}
\concept{Rpath functions}
