% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_usedat}
\alias{test_usedat}
\title{Gulf of Alaska (west/central) 1990s Ecopath model.}
\format{
An Rpath model object
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
Some Source Or Other
}
\usage{
test_usedat
}
\description{
Rpath inputs (unbalanaced model) from the Western/Centeral Gulf of Alaska
1990s Ecopath model.
}
\keyword{datasets}
