% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write.Rsim}
\alias{write.Rsim}
\title{Write function for Rsim run}
\usage{
write.Rsim(Rsim.output, file = NA)
}
\arguments{
\item{Rsim.output}{object created by the \code{rsim.run()} function.}

\item{file}{file name for resultant file.  Need to specify ".csv" or ".RData".}
}
\value{
Writes a list object or file with the start and end biomass and catch per group
from an Rpath.sim object.
\item{Group}{Vector of group names}
\item{StartBio}{Vector of Biomass values at simulation start}
\item{EndBio}{Vector of Biomass values at simulation end}
\item{BioES}{Vector of ratio end:start Biomass values}
\item{StartCatch}{Vector of Catch values at simulation start}
\item{EndCatch}{Vector of Catch values at simulation end}
\item{CatchES}{Vector of ratio end:start Catch values}
}
\description{
Outputs starting biomass, ending biomass, and catch to a .csv or .Rdata file.
}
\examples{
# Read in Rpath parameter file and generate model object
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)
# Run the Rsim simulation
Rsim.output <- rsim.run(Rsim.scenario, method = "RK4", years = 1:50)
# Save summary output to a .csv file
write.Rsim(Rsim.output, file = "output.csv")


}
\seealso{
Other Rsim functions:
\code{\link[=extract.node]{extract.node()}},
\code{\link[=rsim.params]{rsim.params()}},
\code{\link[=rsim.plot]{rsim.plot()}},
\code{\link[=rsim.run]{rsim.run()}},
\code{\link[=rsim.scenario]{rsim.scenario()}},
\code{\link[=rsim.step]{rsim.step()}}
}
\concept{Rsim functions}
