% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AB.params}
\alias{AB.params}
\title{AB.params}
\format{
An Rpath balanced model object that contains:
\describe{
  \item{model}{A data.table containing Ecopath unbalanced ecosystem parameters (base parameters and fisheries).}
  \item{diet}{A data.table containing the Ecopath model's diet matrix.}
  \item{stanzas}{Parameters for multistanza groups.}
  \item{pedigree}{A data.table containing the data quality (pedigree) for the Ecopath model.}   
  ...
}
}
\usage{
AB.params
}
\description{
A tutorial ecosystem for teaching Rpath
}
\keyword{datasets}
