% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{adjust.fishing}
\alias{adjust.fishing}
\title{Adjust Fishing Mortality}
\usage{
adjust.fishing(
  Rsim.scenario,
  parameter,
  group = NA,
  sim.year = 1,
  sim.month = 0,
  value
)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}

\item{parameter}{Parameters to be modified (\code{ForcedEffort},\code{ForcedCatch},\code{ForcedFRate})}

\item{group}{Name of the group whose parameter is being changed. Valid values are found in the `Group` field of the object
created from running \code{rpath()}}

\item{sim.year}{Year of the simulation that should be modified.  Can be a range of years.}

\item{sim.month}{Month of the year that should be modified.  If set to 0, all months of
the year are modified.}

\item{value}{New value for the parameter.}
}
\value{
Returns an \code{Rsim.scenario()} object with the new fishing parameter
   values.
}
\description{
Modifies the fishing mortality value for a species by a particular gear.
Parameters that can be adjusted using this function are: \emph{ForcedEffort},
\emph{ForcedFRate}, or \emph{ForcedCatch}.
}
\examples{
# Read in Rpath parameter file and generate balanced model
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)
# Change value of forcedFRate for Squids in years 3 through 5 to the value of 2 (for all months)
Rsim.scenario.adjusted.fishing <- adjust.fishing(Rsim.scenario, parameter = "ForcedFRate", group = "cod", sim.year = 3:5, value = 2)
head(Rsim.scenario.adjusted.fishing$fishing$ForcedFRate)


}
\seealso{
Other Adjust functions: 
\code{\link{adjust.forcing}()},
\code{\link{adjust.scenario}()},
\code{\link{set.rsim.scene}()}
}
\concept{Adjust functions}
