% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{adjust.fishing}
\alias{adjust.fishing}
\title{Adjust Fishing Mortality}
\usage{
adjust.fishing(
  Rsim.scenario,
  parameter,
  group = NA,
  sim.year = 1,
  sim.month = 0,
  value
)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{rsim.scenario}.}

\item{parameter}{Parameters to be modified (See Description)}

\item{group}{Name of the group whose parameter is being changed.}

\item{sim.year}{Year of the simulation that should be modified.  Can be a range of years.}

\item{sim.month}{Month of the year that should be modified.  If set to 0, all months of
the year are modified.}

\item{value}{New value for the parameter.}
}
\value{
Returns an \code{Rsim.scenario} object with the new fishing parameter 
   values.
}
\description{
Modifies the fishing mortality value for a species by a particular gear.  
Parameters that can be adjusted using this function are: \emph{ForcedEffort},
\emph{ForcedFRate}, or \emph{ForcedCatch}.
}
\seealso{
Other Adjust functions: 
\code{\link{adjust.scenario}()}
}
\concept{Adjust functions}
