% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{adjust.fishing}
\alias{adjust.fishing}
\title{Adjust Fishing Mortality}
\usage{
adjust.fishing(Rsim.scenario, parameter, group = NA, gear = NA, year, value)
}
\arguments{
\item{Rsim.scenario}{Object generated by ecosim.scenario.}

\item{parameter}{One of three parameters to be modified: EFFORT, FRATE, or CATCH}

\item{group}{Name or number of the species group whose fishing parameter is being changed.}

\item{gear}{Name or number of the gear to which the change applies.}

\item{year}{Year of the simulation that should be modified.  Can be a range of years.}

\item{value}{New value for the parameter.}
}
\value{
Returns an Rpath.sim object with the new fishing parameter.
}
\description{
Modifies the fishing mortality value for a species by a particular gear.
}
\seealso{
Other Rpath.functions: \code{\link{adjust.scenario}};
  \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}};
  \code{\link{frate.table}};
  \code{\link{read.rpath.param}};
  \code{\link{rpath.stanzas}}; \code{\link{rpath}};
  \code{\link{rsim.params}}; \code{\link{rsim.plot}};
  \code{\link{rsim.scenario}}; \code{\link{stanzaplot}};
  \code{\link{webplot}}; \code{\link{write.Rpath}};
  \code{\link{write.Rsim}}; \code{\link{write.rpath.param}}
}

