% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{adjust.fishing}
\alias{adjust.fishing}
\title{Adjust Fishing Mortality}
\usage{
adjust.fishing(Rsim.scenario, parameter, group = NA, sim.year, value)
}
\arguments{
\item{Rsim.scenario}{Object generated by ecosim.scenario.}

\item{parameter}{One of three parameters to be modified: EFFORT, FRATE, or CATCH}

\item{group}{Name of the species group whose fishing parameter is being changed.}

\item{sim.year}{Year of the simulation that should be modified.  Can be a range of years.}

\item{value}{New value for the parameter.}
}
\value{
Returns an Rpath.sim object with the new fishing parameter.
}
\description{
Modifies the fishing mortality value for a species by a particular gear.
}
\seealso{
Other Rpath.functions: \code{\link{adjust.forcing}},
  \code{\link{adjust.scenario}},
  \code{\link{check.rpath.params}},
  \code{\link{create.rpath.params}},
  \code{\link{extract.node}}, \code{\link{frate.table}},
  \code{\link{read.rpath.params}},
  \code{\link{rpath.stanzas}}, \code{\link{rpath}},
  \code{\link{rsim.params}}, \code{\link{rsim.plot}},
  \code{\link{rsim.scenario}}, \code{\link{rsim.step}},
  \code{\link{stanzaplot}}, \code{\link{webplot}},
  \code{\link{write.Rpath}}, \code{\link{write.Rsim}},
  \code{\link{write.rpath.params}}
}

