% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{adjust.forcing}
\alias{adjust.forcing}
\title{Adjust Forcing Parameters}
\usage{
adjust.forcing(Rsim.scenario, parameter, group, sim.year = 1, sim.month = 0,
  bymonth = F, value)
}
\arguments{
\item{Rsim.scenario}{Object generated by rsim.scenario.}

\item{parameter}{The Rsim.scenario forcing parameter to be modified.}

\item{group}{The model group that the parameter change will affect.}

\item{sim.year}{Year of the simulation that should be modified.  Can be a range of years.}

\item{sim.month}{Month of the year that should be modified.  If set to 0, all months of
the year are modified.}

\item{bymonth}{Boolean switch to use only the month parameter rather than the 
combination of year and month.  Months begin at 1.}

\item{value}{The new value for the parameter.}
}
\value{
Returns an Rsim.scenario object with the new parameter.
}
\description{
Modifies the various forcing parameters of the rsim scenario object.
}
\seealso{
Other Rpath.functions: \code{\link{adjust.fishing}},
  \code{\link{adjust.scenario}},
  \code{\link{check.rpath.params}},
  \code{\link{create.rpath.params}},
  \code{\link{extract.node}}, \code{\link{frate.table}},
  \code{\link{read.rpath.params}},
  \code{\link{rpath.stanzas}}, \code{\link{rpath}},
  \code{\link{rsim.params}}, \code{\link{rsim.plot}},
  \code{\link{rsim.scenario}}, \code{\link{rsim.step}},
  \code{\link{stanzaplot}}, \code{\link{webplot}},
  \code{\link{write.Rpath}}, \code{\link{write.Rsim}},
  \code{\link{write.rpath.params}}
}

