% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{adjust.scenario}
\alias{adjust.scenario}
\title{Adjust Rsim.scenario parameters}
\usage{
adjust.scenario(Rsim.scenario, parameter, group, groupto = NA, value)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}

\item{parameter}{Parameters to be modified (See Description)}

\item{group}{The model group that the parameter change will affect.  Note that 
a value of \emph{'all'} will affect all groups associate with the groupto
variable.}

\item{groupto}{The corresponding group who's parameter is affecting the group
variable.}

\item{value}{New value for the parameter.}
}
\value{
Returns an \code{Rsim.scenario} object with the new parameter.
}
\description{
Modifies the various parameters of the \code{rsim.scenario} object. Parameters 
that can be adjusted using this function are:
}
\seealso{
Other Adjust functions: 
\code{\link{adjust.fishing}()}
}
\concept{Adjust functions}
