% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{adjust.scenario}
\alias{adjust.scenario}
\title{Adjust Rsim.scenario parameters}
\usage{
adjust.scenario(Rsim.scenario, parameter, group, groupto = NA, value)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}

\item{parameter}{Parameters to be modified (Choose from: \code{B_BaseRef, MzeroMort, 
UnassimRespFrac, ActiveRespFrac, FtimeAdj, FtimeQBOpt, PBopt, NoIntegrate,
HandleSelf, ScrambleSelf, QQ, DD, VV, HandleSwitch, PredPredWeight, PreyPreyWeight})}

\item{group}{The model group that the parameter change will affect.  Note that 
 a value of \emph{'all'} will affect all groups associated with the `groupto`
 variable. Valid values are found in the `Group` field of the object created
from running \code{rpath()}}

\item{groupto}{The corresponding group who's parameter is affecting the group
variable. Required for parameters \code{QQ}, \code{DD}, \code{VV}, \code{HandleSwitch},\code{PredPredWeight},
\code{PreyPreyWeight}}

\item{value}{New value for the parameter.}
}
\value{
Returns an \code{rsim.scenario()} object with the new parameter.
}
\description{
Modifies the various parameters of the \code{rsim.scenario()} object. Parameters that can be adjusted using this function are: 
\emph{B_BaseRef}, \emph{MzeroMort},\emph{UnassimRespFrac}, \emph{ActiveRespFrac}, \emph{FtimeAdj},
\emph{FtimeQBOpt}, \emph{PBopt}, \emph{NoIntegrate},\emph{HandleSelf}, \emph{ScrambleSelf}, \emph{QQ},
\emph{DD}, \emph{VV}, \emph{HandleSwitch}, \emph{PredPredWeight}, \emph{PreyPreyWeight}
}
\examples{
# Read in Rpath parameter file and generate balanced model
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)
# Adjust the PBopt parameter for cod. Set to value = 2
Rsim.scenario.adjusted <- adjust.scenario(Rsim.scenario, parameter="PBopt",group = "cod", groupto = "all", value = 2)   


}
\seealso{
Other Adjust functions: 
\code{\link{adjust.fishing}()},
\code{\link{adjust.forcing}()},
\code{\link{set.rsim.scene}()}
}
\concept{Adjust functions}
