% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{adjust.scenario}
\alias{adjust.scenario}
\title{Adjust Rsim.scenario parameters}
\usage{
adjust.scenario(Rsim.scenario, parameter, group, groupto = NA, value)
}
\arguments{
\item{Rsim.scenario}{Object generated by rsim.scenario.}
}
\value{
Returns an Rsim.scenario object with the new parameter.
}
\description{
Modifies the fishing mortality value for a species by a particular gear.
}
\seealso{
Other Rpath.functions: \code{\link{adjust.fishing}},
  \code{\link{check.rpath.params}},
  \code{\link{create.rpath.params}},
  \code{\link{frate.table}},
  \code{\link{read.rpath.params}},
  \code{\link{rpath.stanzas}}, \code{\link{rpath}},
  \code{\link{rsim.params}}, \code{\link{rsim.plot}},
  \code{\link{rsim.scenario}}, \code{\link{stanzaplot}},
  \code{\link{webplot}}, \code{\link{write.Rpath}},
  \code{\link{write.Rsim}},
  \code{\link{write.rpath.params}}
}

