% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{adjust.scenario}
\alias{adjust.scenario}
\title{Adjust Rsim.scenario parameters}
\usage{
adjust.scenario(Rsim.scenario, parameter, group, groupto = NA, value)
}
\arguments{
\item{Rsim.scenario}{Object generated by rsim.scenario.}

\item{parameter}{The Rsim.scenario parameter to be modified. (Create a table)}

\item{group}{The model group that the parameter change will affect.  Note that 
a value of 'all' will affect all groups associate with the groupto
variable.}

\item{groupto}{The coresponding group who's parameter is affecting the group
variable.}

\item{value}{The new value for the parameter.}
}
\value{
Returns an Rsim.scenario object with the new parameter.
}
\description{
Modifies the various parameters of the rsim scenario object.
}
\seealso{
Other Rpath.functions: \code{\link{adjust.fishing}},
  \code{\link{adjust.forcing}},
  \code{\link{check.rpath.params}},
  \code{\link{create.rpath.params}},
  \code{\link{extract.node}}, \code{\link{frate.table}},
  \code{\link{read.rpath.params}},
  \code{\link{rpath.stanzas}}, \code{\link{rpath}},
  \code{\link{rsim.params}}, \code{\link{rsim.plot}},
  \code{\link{rsim.scenario}}, \code{\link{rsim.step}},
  \code{\link{stanzaplot}}, \code{\link{webplot}},
  \code{\link{write.Rpath}}, \code{\link{write.Rsim}},
  \code{\link{write.rpath.params}}
}

