% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_convert.r
\name{create.rpath.from.eiixml}
\alias{create.rpath.from.eiixml}
\title{Creates an Rpath object from an EwE exported model (EIIXML format)}
\usage{
create.rpath.from.eiixml(eiifile, verbose = FALSE, export = FALSE)
}
\arguments{
\item{eiifile}{Full path to exported EwE XML file}

\item{verbose}{Logical. Use for debugging. If TRUE, prints out useful content}

\item{export}{Logical. Use for debugging. If TRUE, exports the data frames created in the function}
}
\value{
An Rpath object (list) with the following components:
\item{stanzas}{}
\item{pedigree}{}
\item{diet}{}
\item{model}{}
}
\description{
Reads in the eiixml model format (XML) created from exporting a model from Ecobase GUI.
Translates the XML to an rpath object which can then be balanced
}
