% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_convert.R
\name{create.rpath.from.eiixml}
\alias{create.rpath.from.eiixml}
\title{Creates an Rpath object from an EwE exported model (EIIXML format)}
\usage{
create.rpath.from.eiixml(eiifile, verbose = FALSE)
}
\arguments{
\item{eiifile}{Path to exported EwE XML file}

\item{verbose}{Logical. Use for debugging. If TRUE, prints out a list of parsed tables and number of rows read from the XML file}
}
\value{
An Rpath object (list) with the following components:
\item{stanzas}{Stanza parameters}
\item{pedigree}{Pedigree parameters}
\item{diet}{Diet parameters}
\item{model}{Model parameters}
}
\description{
Parses an eiixml file exported using the Ecopath with Ecosim (EwE) GUI
into an unbalanced Rpath model object, which can then be balanced. 
This function was tested on files exported by EwE version 6.7.
}
\examples{
# Import an eiixml file previously exported from the EwE GUI, check that the
# unbalanced model is functional, recalculate age-structured groups based on
# imported stanza parameters (if the imported model has stanzas), and balance
# the model.
eiixml <- system.file("extdata/xml/Western_Bering_Sea.eiixml", package = "Rpath")
rpath_object <- create.rpath.from.eiixml(eiixml)
check.rpath.params(rpath_object)
rpath_object <- rpath.stanzas(rpath_object)
rpath_balanced <- rpath(rpath_object)

}
\seealso{
Other xml: 
\code{\link{import.eiixml}()}
}
\concept{xml}
