% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_convert.r
\name{create.rpath.from.eiixml}
\alias{create.rpath.from.eiixml}
\title{Creates an Rpath object from an EwE exported model (EIIXML format)}
\usage{
create.rpath.from.eiixml(eiifile, verbose = FALSE)
}
\arguments{
\item{eiifile}{Full path to exported EwE XML file}

\item{verbose}{Logical. Use for debugging. If TRUE, prints out useful content}
}
\value{
An Rpath object (list) with the following components:
\item{stanzas}{}
\item{pedigree}{}
\item{diet}{}
\item{model}{}
}
\description{
Reads in the eiixml model format (XML) created from exporting a model from Ecobase GUI.
Translates the XML to an rpath object which can then be balanced
}
\examples{
# Export a model from Ecobase GUI to xml format
# Path to an exported eiixml file
eiixml <- system.file("extdata/xml","West Bering Sea.eiixml", package = "Rpath")
rpath_object <- create.rpath.from.eiixml(eiixml)

}
\seealso{
Other xml: 
\code{\link{import.eiixml}()}
}
\concept{xml}
