% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/param.R
\name{create.rpath.param}
\alias{create.rpath.param}
\title{Create shells for the Rpath parameter files}
\usage{
create.rpath.param(parameter = "model", group = NA, type = NA,
  filename = NA)
}
\arguments{
\item{parameter}{The type of parameter file you are creating.  Choices include "model",
 "diet", "juvenile", and "pedigree".}

\item{group}{Vector of group names.  If parameter equals "juvenile", this should be a vector of
stanza groups only (Juvenile and Adults in one group).}

\item{type}{Numeric vector of group type. Living = 0, Producer = 1, Detritus = 2,
Fleet = 3. Default NA is used for the juvenile and pedigree parameter files.}

\item{filename}{Name of the output file saved as a .csv. If NA the file will not be written.}
}
\value{
Outputs a shell of the parameter file indicated by the parameter variable.  The shell
 is populated with values of NA or logical default values.  Values can then be filled in using
 R or any spreadsheet program.  Use check.rpath.param() to ensure parameter files are filled out
 correctly (NOTE: This does not ensure data is correct just that it is in the right places).
}
\description{
Creates a shell of the parameter files that can then be filled out in R, Excel, or
another spreadsheet program.
}
\seealso{
Other Rpath.functions: \code{\link{Rsim.params}};
  \code{\link{check.rpath.param}}; \code{\link{ecopath}};
  \code{\link{ecosim.plot}}; \code{\link{frate.adjust}};
  \code{\link{frate.table}}; \code{\link{rpath.stanzas}};
  \code{\link{stanzaplot}}; \code{\link{webplot}};
  \code{\link{write.Rpath.sim}}; \code{\link{write.Rpath}}
}

