% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/param.R
\name{create.rpath.param}
\alias{create.rpath.param}
\title{Creates a shell for the Rpath.params object}
\usage{
create.rpath.param(group, type, stgroup = NA, nstanzas = NA)
}
\arguments{
\item{group}{Vector of group names.}

\item{type}{Numeric vector of group type. Living = 0, Producer = 1, Detritus = 2,
Fleet = 3.}

\item{stgroup}{Vector of multistanza group names.  Include NA for non-stanza groups.}

\item{nstanzas}{Numeric vector of the number of stanzas per multistanza groups.}
}
\value{
Outputs a list object of Rpath parameters which are populated with values
 of NA or logical default values.  Values can then be filled in using
 R.  Use check.rpath.param() to ensure parameter files are filled out
 correctly (NOTE: This does not ensure data is correct just that it is
 in the right places).
}
\description{
Creates a shell of the Rpath.params list object which contains the model, diet,
multistanza, and pedigree parameters.
}
\seealso{
Other Rpath.functions: \code{\link{check.rpath.param}};
  \code{\link{frate.adjust}}; \code{\link{frate.table}};
  \code{\link{read.rpath.params}};
  \code{\link{rpath.stanzas}}; \code{\link{rpath}};
  \code{\link{rsim.params}}; \code{\link{rsim.plot}};
  \code{\link{rsim.scenario}}; \code{\link{stanzaplot}};
  \code{\link{webplot}}; \code{\link{write.Rpath.sim}};
  \code{\link{write.Rpath}};
  \code{\link{write.rpath.params}}
}

