% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{create.rpath.params}
\alias{create.rpath.params}
\title{Creates a shell for the Rpath.params object}
\usage{
create.rpath.params(group, type, stgroup = NA)
}
\arguments{
\item{group}{Vector of group names.}

\item{type}{Numeric vector of group type. Living = 0, Producer = 1, Detritus = 2,
Fleet = 3.}

\item{stgroup}{Vector of multistanza group names.  Include NA for non-stanza groups.}
}
\value{
Outputs a list object of \code{Rpath.params} which are populated with values
 of NA or logical default values.  Values can then be filled in using
 R.  Use \code{check.rpath.params()} to ensure parameter files are filled out
 correctly (NOTE: This does not ensure data is correct just that it is
 in the right places).
}
\description{
Creates a shell of the \code{Rpath.params} list object which contains the \emph{model}, \emph{diet},
\emph{multistanza}, and \emph{pedigree} parameters.
}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{check.rpath.params}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath}()},
\code{\link{rpath.consumers}()},
\code{\link{rpath.detrital}()},
\code{\link{rpath.gears}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.living}()},
\code{\link{rpath.mixotrophs}()},
\code{\link{rpath.producers}()},
\code{\link{rpath.stanzas}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
