% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecosim_plot.R
\name{ecosim.plot}
\alias{ecosim.plot}
\title{Plot routine for Ecosim runs}
\usage{
ecosim.plot(Rpath.sim.obj)
}
\arguments{
\item{Rpath.sim.obj}{Rpath ecosim run created by the ecosim.run() function.}
}
\value{
Creates a figure of relative biomass.
}
\description{
Plots the relative biomass of each group from a run of ecosim.
}
\seealso{
Other Rpath.functions: \code{\link{Rsim.params}};
  \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}}; \code{\link{ecopath}};
  \code{\link{frate.adjust}}; \code{\link{frate.table}};
  \code{\link{rpath.stanzas}}; \code{\link{webplot}};
  \code{\link{write.Rpath.sim}}; \code{\link{write.Rpath}}
}

