% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.R
\name{extract.node}
\alias{extract.node}
\title{Extract node data}
\usage{
extract.node(Rsim.output, group)
}
\arguments{
\item{Rsim.output}{Object generated by [rsim.run()].}

\item{group}{Name of the node for which you want to extract data.}
}
\value{
Returns a list object with node specific data.
}
\description{
Creates a list object with node specific data using output from rsim.run()
}
\examples{
\dontrun{
Rpath <- rpath(Ecosense.EBS)
rsim.scenario(Rpath, Rpath.params, years = 1:100)
Rsim.output <- rsim.run(Rsim.scenario, method = "RK4", years = 1:100)
extract.node(Rsim.output)
}
}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{adjust.forcing}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.stanzas}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.Rsim}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
