% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.R
\name{extract.node}
\alias{extract.node}
\title{Extract node data}
\usage{
extract.node(Rsim.output, group)
}
\arguments{
\item{Rsim.output}{Object generated by rsim.run.}

\item{group}{Name of the node for which you want to extract data.}
}
\value{
Returns a list object with node specific data.
}
\description{
Creates a list object with node specific data using Rsim.output object.
}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{adjust.forcing}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath}()},
\code{\link{rpath.consumers}()},
\code{\link{rpath.detrital}()},
\code{\link{rpath.gears}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.living}()},
\code{\link{rpath.mixotrophs}()},
\code{\link{rpath.producers}()},
\code{\link{rpath.stanzas}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.Rsim}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
