% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{frate.table}
\alias{frate.table}
\title{Fishing Mortality Table}
\usage{
frate.table(Rsim.scenario)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}
}
\value{
Returns a data table of F values for each species/gear combination.
\item{Group}{Group/node}
\item{Fishery/gearType}{Gear type for the fishery} ...
\item{Total}{Sum over all gear types}
}
\description{
Creates a table of fishing mortalities by species group and gear for an 
\code{rsim.scenario()} object.
}
\examples{
# Read in Rpath parameter file and generate balanced model
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)
rates <- frate.table(Rsim.scenario)
# display the head of the data frame
head(rates)

}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath}()},
\code{\link{rpath.consumers}()},
\code{\link{rpath.detrital}()},
\code{\link{rpath.gears}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.living}()},
\code{\link{rpath.mixotrophs}()},
\code{\link{rpath.producers}()},
\code{\link{rpath.stanzas}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
