% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{frate.table}
\alias{frate.table}
\title{Fishing Mortality Table}
\usage{
frate.table(Rsim.scenario)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{rsim.scenario}.}
}
\value{
Returns a data table of F values for each species/gear combination.
}
\description{
Creates a table of fishing mortalities by species group and gear for an 
\code{Rsim.scenario} object.
}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{adjust.forcing}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{extract.node}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.stanzas}()},
\code{\link{rpath}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.Rsim}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
