% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{frate.table}
\alias{frate.table}
\title{Fishing Mortality Table}
\usage{
frate.table(Rpath.sim)
}
\arguments{
\item{Rpath.sim}{Object generated by ecosim.init.}
}
\value{
Returns a data table of F values for each species/gear combination.
}
\description{
Creates a table of fishing mortalities by species group and gear for an Rpath.sim object.
}
\seealso{
Other Rpath.functions: \code{\link{Rsim.params}};
  \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}}; \code{\link{ecopath}};
  \code{\link{ecosim.plot}}; \code{\link{frate.adjust}};
  \code{\link{webplot}}; \code{\link{write.Rpath.sim}};
  \code{\link{write.Rpath}}
}

