% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{frate.table}
\alias{frate.table}
\title{Fishing Mortality Table}
\usage{
frate.table(Rsim.scenario)
}
\arguments{
\item{Rsim.scenario}{Object generated by rsim.scenario.}
}
\value{
Returns a data table of F values for each species/gear combination.
}
\description{
Creates a table of fishing mortalities by species group and gear for an Rpath.sim object.
}
\seealso{
Other Rpath.functions: \code{\link{adjust.fishing}},
  \code{\link{adjust.forcing}},
  \code{\link{adjust.scenario}},
  \code{\link{check.rpath.params}},
  \code{\link{create.rpath.params}},
  \code{\link{extract.node}},
  \code{\link{read.rpath.params}},
  \code{\link{rpath.stanzas}}, \code{\link{rpath}},
  \code{\link{rsim.params}}, \code{\link{rsim.plot}},
  \code{\link{rsim.scenario}}, \code{\link{rsim.step}},
  \code{\link{stanzaplot}}, \code{\link{webplot}},
  \code{\link{write.Rpath}}, \code{\link{write.Rsim}},
  \code{\link{write.rpath.params}}
}

