% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{get.rsim.fishing}
\alias{get.rsim.fishing}
\title{Retrieve Rsim fishing forcing parameters}
\usage{
get.rsim.fishing(Rsim.scenario)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{rsim.scenario}.}
}
\value{
Returns an \code{Rsim.fishing} object.
}
\description{
Helper function that will retrieve the fishing forcing parameters that were used in an 
Rsim scenario
}
\seealso{
Other Get functions: 
\code{\link{get.rsim.forcing}()},
\code{\link{get.rsim.params}()},
\code{\link{get.rsim.stanzas}()},
\code{\link{get.rsim.start_state}()}
}
\concept{Get functions}
