% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{get.rsim.params}
\alias{get.rsim.params}
\title{Retrieve Rsim scenario parameters}
\usage{
get.rsim.params(Rsim.scenario)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}
}
\value{
Returns an \code{Rsim.params} object.
}
\description{
Helper function that will retrieve parameters that were used in an Rsim scenario
}
\seealso{
Other Get functions: 
\code{\link{get.rsim.fishing}()},
\code{\link{get.rsim.forcing}()},
\code{\link{get.rsim.stanzas}()},
\code{\link{get.rsim.start_state}()}
}
\concept{Get functions}
