% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{get.rsim.stanzas}
\alias{get.rsim.stanzas}
\title{Retrieve Rsim stanza parameters}
\usage{
get.rsim.stanzas(Rsim.scenario)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}
}
\value{
Returns an \code{Rsim.stanzas} object.
}
\description{
Helper function that will retrieve the stanza parameters that were used in an 
Rsim scenario
}
\seealso{
Other Get functions: 
\code{\link{get.rsim.fishing}()},
\code{\link{get.rsim.forcing}()},
\code{\link{get.rsim.params}()},
\code{\link{get.rsim.start_state}()}
}
\concept{Get functions}
