% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{get.rsim.start_state}
\alias{get.rsim.start_state}
\title{Retrieve Rsim starting state values}
\usage{
get.rsim.start_state(Rsim.scenario)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{\link{rsim.scenario}()}.}
}
\value{
Returns an \code{Rsim.state} object.
}
\description{
Helper function that will retrieve the starting state values that were used 
in an Rsim scenario
}
\seealso{
Other Get functions: 
\code{\link{get.rsim.fishing}()},
\code{\link{get.rsim.forcing}()},
\code{\link{get.rsim.params}()},
\code{\link{get.rsim.stanzas}()}
}
\concept{Get functions}
