% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{get.rsim.start_state}
\alias{get.rsim.start_state}
\title{Retrieve starting state values from an Rsim scenario}
\usage{
get.rsim.start_state(Rsim.scenario)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{rsim.scenario}.}
}
\value{
Returns an `start_state` object.
}
\description{
Helper function that will retrieve the starting state values that were used 
in an Rsim scenario
}
\examples{
# Read in Rpath parameter file and balance model
Rpath <- rpath(AB.params)
# Create a 50 yr Rpath scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)
params <- get.rsim.start_state(Rsim.scenario)
names(params) 



}
\seealso{
Other Get functions: 
\code{\link{get.rsim.fishing}()},
\code{\link{get.rsim.forcing}()},
\code{\link{get.rsim.params}()},
\code{\link{get.rsim.stanzas}()}
}
\concept{Get functions}
