% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopath_plot.R
\name{ggwebplot}
\alias{ggwebplot}
\title{ggplot routine for Rpath food web}
\usage{
ggwebplot(
  Rpath.obj,
  eco.name = attr(Rpath.obj, "eco.name"),
  line.col = "grey",
  highlight = NULL,
  highlight.col = c("black", "red", "orange"),
  labels = FALSE,
  label.num = FALSE,
  label.cex = 1,
  fleets = FALSE,
  type.col = "grey50",
  box.order = NULL,
  line.alpha = 0.5,
  max.overlaps = 10
)
}
\arguments{
\item{Rpath.obj}{Rpath model created by the `rpath()` function.}

\item{eco.name}{Optional name of the ecosystem. Default is the `eco.name` 
attribute from the rpath object.}

\item{line.col}{The color of the lines between nodes of the food web.}

\item{highlight}{Set to the group number or name to highlight the connections of that group.}

\item{highlight.col}{Color of the connections to the highlighted group, vector of length 3}

\item{labels}{Logical whether or not to display group names.}

\item{label.num}{Logical whether or not to display group numbers instead of points at nodes. 
If `TRUE`, `type.col` must be length 1, not 4.}

\item{label.cex}{}

\item{fleets}{Logical value indicating whether or not to include fishing fleets in the food web.}

\item{type.col}{The color of the points cooresponding to the types of the group.  Can either be 
of length 1 or 4. Color order will be living, primary producers, detrital, and fleet groups.}

\item{box.order}{Vector of box numbers to change the default plot order. Must include all box numbers. 
Passed to `summarize.for.webplot()`}

\item{line.alpha}{Transparency of lines between nodes of the food web}

\item{max.overlaps}{Maximum number of overlaps allowed for group labels by `ggrepel`
Defaults to 10.}
}
\value{
Returns a ggplot object visualizing the food web
}
\description{
Plots the food web associated with an Rpath object using ggplot functions
}
