% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_convert.r
\name{import.eiixml}
\alias{import.eiixml}
\title{Reads in EwE exported XML file and parses into data frames}
\usage{
import.eiixml(eiifile, verbose = F, export = F)
}
\arguments{
\item{eiifile}{Full path to exported EwE XML file}

\item{verbose}{Logical. Use for debugging. If TRUE, prints out useful content}

\item{export}{Logical. Use for debugging. If TRUE, exports the data frames created in the function}
}
\value{
A list of data frames, one for each table in the XML file.
}
\description{
Parses the exported eiixml file (XML format, exported using Ecobase GUI) file into a list of data frames, one for each table in the XML.
To be useful in Rpath this needs to be further processed into an Rpath object.
}
