% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_functions.R
\name{MTI}
\alias{MTI}
\title{Mixed Trophic Impacts}
\usage{
MTI(Rpath, Rpath.params, increase = T)
}
\arguments{
\item{Rpath}{Balanced Rpath model generated by rpath.}

\item{Rpath.params}{R object containing the Rpath parameters.  This is generated
either by the create.rpath.params or read.rpath.params functions.}

\item{increase}{Logical value indicating whether a marginal increase is applied.}
}
\value{
Returns a matrix of mixed trophic impacts.
}
\description{
Need to edit.
}
\seealso{
Other Rpath functions: 
\code{\link{adjust.forcing}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{extract.node}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath}()},
\code{\link{rpath.consumers}()},
\code{\link{rpath.detrital}()},
\code{\link{rpath.gears}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.living}()},
\code{\link{rpath.mixotrophs}()},
\code{\link{rpath.producers}()},
\code{\link{rpath.stanzas}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.Rsim}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
