% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpath_pkg.R
\docType{package}
\name{Rpath-package}
\alias{Rpath-package}
\title{Rpath: A package implementating mass balance algorithms designed to work with 
fisheries data sources.}
\description{
The Rpath package provides two categories of important functions:
rpath and rsim.
}
\section{rpath functions}{

The rpath functions generate the balanced snap shot of energy flow through the
system.  This is done by solving a series of linear equations for unknown biomass
or ecotrophic efficiency.
}

