% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpath-package.R, R/Rpath_pkg.R
\docType{package}
\name{Rpath-package}
\alias{Rpath-package}
\title{Rpath: A package implementing mass balance algorithms designed to work with 
fisheries data sources.}
\description{
The Rpath package provides two categories of important functions:
rpath and rsim.

The Rpath package provides two categories of important functions:
rpath and rsim.
}
\section{Rpath functions}{

The Rpath functions generate the balanced snap shot of energy flow through the
system.  This is done by solving a series of linear equations for unknown biomass
or ecotrophic efficiency.
}

\section{rpath functions}{

The rpath functions generate the balanced snap shot of energy flow through the
system.  This is done by solving a series of linear equations for unknown biomass
or ecotrophic efficiency.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://noaa-edab.github.io/Rpath/,https://github.com/NOAA-EDAB/Rpath/}
  \item Report bugs at \url{https://github.com/NOAA-EDAB/Rpath/issues}
}


Useful links:
\itemize{
  \item \url{https://noaa-edab.github.io/Rpath/,https://github.com/NOAA-EDAB/Rpath/}
  \item Report bugs at \url{https://github.com/NOAA-EDAB/Rpath/issues}
}

}
\author{
\strong{Maintainer}: Kerim Aydin \email{kerim.aydin@noaa.gov} (\href{https://orcid.org/0000-0003-3792-9828}{ORCID})

Authors:
\itemize{
  \item Sean Lucey \email{sean.m.lucey@noaa.gov} (\href{https://orcid.org/0000-0001-6675-5500}{ORCID})
  \item Sarah Gaichas \email{sgaichas@hydrascientificllc.com} (\href{https://orcid.org/0000-0002-5788-3073}{ORCID})
  \item Sarah Weisberg \email{sarahjweisberg@gmail.com} (\href{https://orcid.org/0000-0003-4271-9534}{ORCID})
  \item Andy Whitehouse \email{gaw@uw.edu} (\href{https://orcid.org/0000-0002-9130-9403}{ORCID})
}

Other contributors:
\itemize{
  \item Bia Dias \email{bia.dias@noaa.gov} (\href{https://orcid.org/0000-0001-5905-4044}{ORCID}) [contributor]
  \item Ron Klasky \email{ronald.klasky@noaa.gov} [contributor]
  \item Andy Beet \email{andrew.beet@noaa.gov} (\href{https://orcid.org/0000-0001-8270-7090}{ORCID}) [contributor]
  \item Max Grezlik \email{maxwell.grezlik@noaa.gov} [contributor]
}

}
