% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpath_support.R
\name{rpath.consumers}
\alias{rpath.consumers}
\title{List of Rpath consumer groups}
\usage{
rpath.consumers(Rpath)
}
\arguments{
\item{Rpath}{Balanced Rpath model generated by \code{rpath()}}
}
\value{
Returns a string vector containing the names of Rpath functional groups
that are consumers
}
\description{
List of consumer groups from Rpath object with species type = 0.
}
\examples{
# Build the balanced Rpath model and parameter file by calling `rpath`
Rpath <- rpath(Ecosense.EBS)
# Run the consumers command
consumerGroups = rpath.consumers(Rpath)
# Print out the first few consumer group names
head(consumerGroups)

}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{adjust.forcing}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath}()},
\code{\link{rpath.detrital}()},
\code{\link{rpath.gears}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.living}()},
\code{\link{rpath.mixotrophs}()},
\code{\link{rpath.producers}()},
\code{\link{rpath.stanzas}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
