% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpath_support.R
\name{rpath.groups}
\alias{rpath.groups}
\alias{rpath.living}
\alias{rpath.detrital}
\alias{rpath.gears}
\alias{rpath.producers}
\alias{rpath.consumers}
\alias{rpath.mixotrophs}
\title{Rpath functional group names}
\usage{
rpath.groups(Rpath)

rpath.living(Rpath)

rpath.detrital(Rpath)

rpath.gears(Rpath)

rpath.producers(Rpath)

rpath.consumers(Rpath)

rpath.mixotrophs(Rpath)
}
\arguments{
\item{Rpath}{Balanced Rpath model generated by rpath.}
}
\value{
Returns a character vector containing the names of Rpath functional groups
by category (group type).
}
\description{
Get a character vector of functional group names from an Rpath object (balanced model)  
or an Rpath.params object (unbalanced model parameters) based on the 'type' input
parameter as follows: (0: consumers, 1: producers, 2: detrital, 3: gears, 0<type<1: mixotrophs).
Living groups are consumers + producers.  Note that mixotrophs are not returned as
either consumers or producers, only separately.
}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{adjust.forcing}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{extract.node}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath}()},
\code{\link{rpath.stanzas}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.Rsim}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
