% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopath.R
\name{rpath}
\alias{rpath}
\title{Ecopath module of Rpath}
\usage{
rpath(Rpath.params, eco.name = NA, eco.area = 1)
}
\arguments{
\item{Rpath.params}{R object containing the parameters needed to create a
\code{Rpath} model.  This is generated either by the \code{\link{create.rpath.params}()}
 or \code{\link{read.rpath.params}()} functions.}

\item{eco.name}{Optional name of the ecosystem which becomes an attribute of the
\code{Rpath} model.}

\item{eco.area}{Optional area of the ecosystem which becomes an attribute of the
\code{Rpath} model.}
}
\value{
Returns a static \code{Rpath} model that can be supplied to the
\code{rsim.scenario} function.
}
\description{
Performs initial mass balance using a \code{Rpath.params()} file
}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath.consumers}()},
\code{\link{rpath.detrital}()},
\code{\link{rpath.gears}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.living}()},
\code{\link{rpath.mixotrophs}()},
\code{\link{rpath.producers}()},
\code{\link{rpath.stanzas}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
