% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecopath.R
\name{rpath}
\alias{rpath}
\title{Ecopath modual of Rpath}
\usage{
rpath(modfile, dietfile, eco.name = NA)
}
\arguments{
\item{modfile}{Comma deliminated model parameter file.}

\item{dietfile}{Comma deliminated diet matrix file.}

\item{eco.name}{Optional name of the ecosystem which becomes an attribute of
   rpath object.}
}
\value{
Returns an Rpath object that can be supplied to the ecosim.init function.
}
\description{
Performs initial mass balance using a model parameter file and diet
matrix file.
}
\seealso{
Other Rpath.functions: \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}};
  \code{\link{ecosim.plot}}; \code{\link{frate.adjust}};
  \code{\link{frate.table}}; \code{\link{rpath.stanzas}};
  \code{\link{rsim.params}}; \code{\link{stanzaplot}};
  \code{\link{webplot}}; \code{\link{write.Rpath.sim}};
  \code{\link{write.Rpath}}
}

