% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopath.R
\name{rpath.stanzas}
\alias{rpath.stanzas}
\title{Calculate biomass and consumption for multistanza groups}
\usage{
rpath.stanzas(Rpath.params)
}
\arguments{
\item{Rpath.params}{R object containing the parameters needed to create a
\code{Rpath} model.  This is generated either by the \code{\link{create.rpath.params}()}
 or \code{\link{read.rpath.params}()} functions.}
}
\value{
Calculates and adds biomass and consumption for trailing stanza groups.
 Also adds weight at age and number at age for multi-staza groups.
}
\description{
Uses the leading stanza to calculate the biomass and consumption of other stanzas
necessary to support the leading stanza.
}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath}()},
\code{\link{rpath.consumers}()},
\code{\link{rpath.detrital}()},
\code{\link{rpath.gears}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.living}()},
\code{\link{rpath.mixotrophs}()},
\code{\link{rpath.producers}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
