% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecopath.R
\name{rpath.stanzas}
\alias{rpath.stanzas}
\title{Calculate biomass and consumption for multistanza groups}
\usage{
rpath.stanzas(modfile, juvfile)
}
\arguments{
\item{modfile}{Object containing the model parameters.}

\item{juvfile}{Object containing the group and stanza specific characteristics.}
}
\value{
Adds the biomass and consumption to the relative groups in the model parameter
 object.  Need to assign the output to the juvfile if you want to plot stanza group
 relative biomass.
}
\description{
Uses the leading stanza to calculate the biomass and consumption of other stanzas
necessary to support the leading stanza.
}
\seealso{
Other Rpath.functions: \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}}; \code{\link{ecopath}};
  \code{\link{ecosim.plot}}; \code{\link{frate.adjust}};
  \code{\link{frate.table}}; \code{\link{rsim.params}};
  \code{\link{stanzaplot}}; \code{\link{webplot}};
  \code{\link{write.Rpath.sim}}; \code{\link{write.Rpath}}
}

