% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecopath.R
\name{rpath.stanzas}
\alias{rpath.stanzas}
\title{Calculate biomass and consumption for multistanza groups}
\usage{
rpath.stanzas(Rpath, groupfile, stanzafile)
}
\arguments{
\item{groupfile}{Object containing the group specific characteristics.}

\item{stanzafile}{Object containing the stanza specific characteristics.}

\item{Rpath.obj}{Rpath model created by the ecopath() function.}
}
\value{
Adds the biomass and consumption to the relative groups in the model parameter
object
}
\description{
Uses the leading stanza to calculate the biomass and consumption of other stanzas
necessary to support the leading stanza.
}
\seealso{
Other Rpath.functions: \code{\link{Rsim.params}};
  \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}}; \code{\link{ecopath}};
  \code{\link{ecosim.plot}}; \code{\link{frate.adjust}};
  \code{\link{frate.table}}; \code{\link{webplot}};
  \code{\link{write.Rpath.sim}}; \code{\link{write.Rpath}}
}

