% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.deriv}
\alias{rsim.deriv}
\title{Calculate the derivatives for a time step}
\usage{
rsim.deriv(Rsim.scenario, sim.year = 0, sim.month = 0, tstep = 0)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{rsim.scenario}.}

\item{sim.year}{Will inherit from apply functions}

\item{sim.month}{Will inherit from apply functions}

\item{tstep}{Sub-monthly time step usually set to 0.}
}
\description{
Calculates the derivative for a single time step and saves the output
}
