% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.fishing}
\alias{rsim.fishing}
\title{Generate Rsim fishing matrix}
\usage{
rsim.fishing(params, years)
}
\arguments{
\item{params}{Rsim parameter object generated by \code{rsim.params}.}

\item{years}{A vector of each year of the simulation.}
}
\description{
Creates a matrix for forcing functions related to fishing (\emph{ForcedEffort}, 
\emph{ForcedFRate}, \emph{ForcedCatch})
}
