% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.forcing}
\alias{rsim.forcing}
\title{Generate Rsim forcing matrix}
\usage{
rsim.forcing(params, years)
}
\arguments{
\item{params}{Rsim parameter object generated by \code{rsim.params}.}

\item{years}{A vector of each year of the simulation.}
}
\description{
Creates a matrix for forcing functions not related to fishing (\emph{ForcedPrey}
\emph{ForcedMort}, \emph{ForcedRecs}, \emph{ForcedSearch}, \emph{ForcedMigrate}, 
\emph{ForcedBio}).
}
