% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.mort}
\alias{rsim.mort}
\title{Output mortality on a group}
\usage{
rsim.mort(Rsim.output, group)
}
\arguments{
\item{Rsim.output}{R object containing the output from \code{rsim.run}.}

\item{group}{Group from the \code{Rpath} model that is of interest}
}
\description{
Creates a matrix of mortality by predators on a particular prey.
}
