% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim_plot.R
\name{rsim.plot}
\alias{rsim.plot}
\title{Plot routine for Rsim runs}
\usage{
rsim.plot(Rsim.output, spname = "all", indplot = F)
}
\arguments{
\item{Rsim.output}{R object containing the output from \code{rsim.run}.}

\item{spname}{Vector of species names to be displayed.  The default "all" will
display all group names.}

\item{indplot}{Logical value of whether to plot a single group or multiple groups}
}
\value{
Creates a figure of relative biomass.
}
\description{
Plots the relative biomass of each group from an \code{Rsim.output} object.
}
\seealso{
Other Rsim functions: 
\code{\link{rsim.params}()},
\code{\link{rsim.run}()},
\code{\link{rsim.scenario}()},
\code{\link{rsim.step}()}
}
\concept{Rsim functions}
