% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim_plot.R
\name{rsim.plot}
\alias{rsim.plot}
\title{Plot routine for Ecosim runs}
\usage{
rsim.plot(Rsim.output, spname)
}
\arguments{
\item{Rsim.output}{Rpath ecosim run created by the rsim.run() function.}
}
\value{
Creates a figure of relative biomass.
}
\description{
Plots the relative biomass of each group from a run of ecosim.
}
\seealso{
Other Rpath.functions: \code{\link{adjust.fishing}},
  \code{\link{adjust.scenario}},
  \code{\link{check.rpath.param}},
  \code{\link{create.rpath.param}},
  \code{\link{frate.table}},
  \code{\link{read.rpath.params}},
  \code{\link{rpath.stanzas}}, \code{\link{rpath}},
  \code{\link{rsim.params}}, \code{\link{rsim.scenario}},
  \code{\link{stanzaplot}}, \code{\link{webplot}},
  \code{\link{write.Rpath.sim}}, \code{\link{write.Rpath}},
  \code{\link{write.rpath.params}}
}

