% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.run}
\alias{rsim.run}
\title{Run Rsim}
\usage{
rsim.run(Rsim.scenario, method = "RK4", years = 1:100)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{rsim.scenario}.}

\item{method}{Numerical integration method. Either \emph{'AB'} for Adams-Bashforth or 
\emph{'RK4'} for 4th order Runge-Kutta.}

\item{years}{A vector of each year of the simulation.}
}
\description{
Carries out the numerical integration of the Rsim alogrithms.
}
\seealso{
Other Rsim functions: 
\code{\link{rsim.params}()},
\code{\link{rsim.plot}()},
\code{\link{rsim.scenario}()},
\code{\link{rsim.step}()}
}
\concept{Rsim functions}
