% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.scenario}
\alias{rsim.scenario}
\title{Rsim module of \code{Rpath}}
\usage{
rsim.scenario(Rpath, Rpath.params, years = 1:100)
}
\arguments{
\item{Rpath}{R object containing a balanced \code{Rpath} model.}

\item{Rpath.params}{R object containing the Rpath parameters.  This is generated
either by the create.rpath.params or read.rpath.params functions.}

\item{years}{A vector of each year of the simulation.}
}
\value{
Returns an \code{Rsim.scenario} object that can be supplied to the 
    \code{rsim.run} function.
}
\description{
Uses a balanced \code{Rpath} model and creates a scenario consisting of 5 list objects:
\code{params}, \code{start_state}, \code{forcing}, \code{fishing}, and 
\code{stanzas}.
}
\seealso{
Other Rsim functions: 
\code{\link{extract.node}()},
\code{\link{rsim.params}()},
\code{\link{rsim.plot}()},
\code{\link{rsim.run}()},
\code{\link{rsim.step}()},
\code{\link{write.Rsim}()}
}
\concept{Rsim functions}
