% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosense.R
\name{rsim.sense}
\alias{rsim.sense}
\title{Ecosense function for rpath (rsim.sense)}
\usage{
rsim.sense(Rsim.scenario, Rpath.params, Vvary = c(0, 0), Dvary = c(0, 0))
}
\arguments{
\item{Rsim.scenario}{Rsim scenario object used as center of distributions
(base model) during random parameter generation.}

\item{Rpath.params}{Rpath parameter object (unbalanced model) used for
data pedigree input.}

\item{Vvary}{length-2 vector with (lower,upper) bounds of vulnerability
generation in log-space - 1, Vvary = log(X-1) so Walters et al. 1997
range of (1..inf) centered on 2 becomes (-inf,+inf) centered on 0.}

\item{Dvary}{length-2 vector with (lower,upper) bounds of handling time
generation in log-space - 1, scaled as Vvary, above.}
}
\value{
Returns an Rsim.params object that can be substituted for the params
   in an Rsim.scenario object.
}
\description{
Kerim Aydin 31 December 2019
}
\concept{Rpath functions

This function generates random parameters around a given rsim scenario
object.  

It has been tested to give the same results as Whitehouse and Aydin 
(Submitted) when supplied with a straight-from-ecopath rsim secnario.
and the same random seed - though some minor bug corrections have been
found in the previous version (previous version is rsim.sense.path)}
