% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.stanzas}
\alias{rsim.stanzas}
\title{Generate Rsim stanza matrix}
\usage{
rsim.stanzas(Rpath.params, state, params)
}
\arguments{
\item{Rpath.params}{R object containing the Rpath parameters.  This is generated
either by the create.rpath.params or read.rpath.params functions.}

\item{state}{List object of state variables generated by \code{rsim.state}.}

\item{params}{Rsim parameter object generated by \code{rsim.params}.}
}
\description{
Creates a matrix of stanza variables to be used by \code{rsim.run}.
}
