% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.stanzas}
\alias{rsim.stanzas}
\title{Initialize list of age structured group (stanza) parameters}
\usage{
rsim.stanzas(Rpath.params, state, params)
}
\arguments{
\item{Rpath.params}{R object containing the parameters needed to create a
\code{Rpath} model.  This is generated either by the \code{\link{create.rpath.params}()}
 or \code{\link{read.rpath.params}()} functions.}

\item{state}{List object of state variables generated by \code{rsim.state}.}

\item{params}{Rsim parameter object generated by \code{rsim.params}.}
}
\value{
a list of 21 objects:
\itemize{
 \item{\code{Nsplit}, number of model groups with stanzas}
 \item{\code{Nstanzas}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by the number of age groups for each model group with stanzas}
 \item{\code{EcopathCode}, numeric matrix of Rpath group numbers with \code{Nsplit}+1 species rows and \code{Nstanzas}+1} age group columns, first column and row empty for indexing}
 \item{\code{Age1}, numeric matrix of age in months of first age group with \code{Nsplit}+1 species rows, first column and row empty for indexing, second column earliest age in months (0), third column max age in months
 \item{\code{Age2}, numeric matrix of age in months of last age group with \code{Nsplit}+1 species rows, first column and row empty for indexing, second column earliest age in months, third column max age in months}
 \item{\code{baseWageS}, numeric matrix}
 \item{\code{baseNageS}, numeric matrix}
 \item{\code{baseQageS}, numeric matrix}
 \item{\code{Wmat}, numeric vector}
 \item{\code{RecPower}, numeric vector}
 \item{\code{recruits}, numeric vector}
 \item{\code{VBGFd}, numeric vector}
 \item{\code{RzeroS}, numeric vector}
 \item{\code{vBM}, numeric vector}
 \item{\code{baseEggsStanza}, numeric vector}
 \item{\code{SplitAlpha}, numeric matrix}
 \item{\code{SpawnX}, numeric vector}
 \item{\code{SpawnEnergy}, numeric vector}
 \item{\code{baseSpawnBio}, numeric vector of initial age structured stanza }
 \item{\code{StanzaPred}, numeric vector}
 \item{\code{RscaleSplit}, numeric vector}
 \item{\code{baseStanzaPred}, numeric matrix of monthly (rows) biomass multiplier by biomass group (columns)}
}
}
\description{
Creates a list of stanza indices and parameters to be used by \code{rsim.run}.
}
\examples{
# Read in Rpath parameter file and generate model object
Rpath <- rpath(REco.params)
# Create default dynamic parameters from Rpath model
Rsim.params <- rsim.params(Rpath)
# Create state vectors with default values
Rsim.state <- rsim.state(Rsim.params)
# Create stanza object
Rsim.stanzas <- rsim.stanzas(REco.params, Rsim.state, Rsim.params)

}
