% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.stanzas}
\alias{rsim.stanzas}
\title{Initialize list of age structured group (stanza) parameters for dynamic simulation}
\usage{
rsim.stanzas(Rpath.params, state, params)
}
\arguments{
\item{Rpath.params}{R object containing the parameters needed to create a
\code{Rpath} model.  This is generated either by the \code{\link{create.rpath.params}()}
 or \code{\link{read.rpath.params}()} functions.}

\item{state}{List object of state variables generated by \code{rsim.state}.}

\item{params}{Rsim parameter object generated by \code{rsim.params}.}
}
\value{
a list of 21 objects:
\itemize{
 \item{\code{Nsplit}, number of model groups with stanzas}
 \item{\code{Nstanzas}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by the number of age groups for each model group with stanzas}
 \item{\code{EcopathCode}, numeric matrix of Rpath group numbers with \code{Nsplit}+1 species rows and \code{Nstanzas}+1 age group columns, first column and row empty for indexing}
 \item{\code{Age1}, numeric matrix of age in months of first age group with \code{Nsplit}+1 species rows, first column and row empty for indexing, second column earliest age in months (0), third column max age in months}
 \item{\code{Age2}, numeric matrix of age in months of last age group with \code{Nsplit}+1 species rows, first column and row empty for indexing, second column earliest age in months, third column max age in months}
 \item{\code{baseWageS}, numeric matrix with rows of monthly weight relative to weight at infinity (Winf) calculated in \code{rpath.stanzas} with \code{Nsplit}+1 species columns}
 \item{\code{baseNageS}, numeric matrix with rows of monthly numbers surviving, calculated in \code{rpath.stanzas} with \code{Nsplit}+1 species columns}
 \item{\code{baseQageS}, numeric matrix with rows of monthly consumption, calculated as WageS^VBGFd in \code{rpath.stanzas} with \code{Nsplit}+1 species columns}
 \item{\code{Wmat}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by weight at 50\% maturity relative to Winf for each model group with stanzas}
 \item{\code{RecPower}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by recruitment power parameter for each model group with stanzas}
 \item{\code{recruits}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by recruits per spawner entered as input to \code{rpath.stanzas}}
 \item{\code{VBGFd}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by von Bertalanffy d (default = 2/3) for each model group with stanzas}
 \item{\code{RzeroS}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by recruits per spawner entered as input to \code{rpath.stanzas}}
 \item{\code{vBM}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by von Bertalanffy growth coefficient for each model group with stanzas}
 \item{\code{SplitAlpha}, numeric matrix of initial monthly (rows) growth coefficients for each model group with stanzas (\code{Nsplit}+1 species columns)}
 \item{\code{SpawnX}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by relative strength of spawner-recruit relationship for each model group with stanzas, currently set to 10000, no relationship}
 \item{\code{SpawnEnergy}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by relative energy going to recruitment rather than growth for each model group with stanzas, currently set to 1}
 \item{\code{baseEggsStanza}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by initial estimate of egg output for each model group with stanzas}
 \item{\code{baseSpawnBio}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by initial estimate of spawning biomass (=egg output) for each model group with stanzas}
 \item{\code{RscaleSplit}, numeric vector length \code{Nsplit}+1, leading 0 for indexing followed by adult to juvenile metric scaling parameter for each model group with stanzas, currently set to 1}
 \item{\code{baseStanzaPred}, numeric vector length \code{NUM_GROUPS}+1, base consumption (sum of QageS*NageS) for each model group with stanzas in full model group order; 0s entered for non-stanza groups}
}
}
\description{
Creates a list of stanza indices and parameters to be used by \code{rsim.run}.
}
\examples{
# Read in Rpath parameter file and generate model object
Rpath <- rpath(REco.params)
# Create default dynamic parameters from Rpath model
Rsim.params <- rsim.params(Rpath)
# Create state vectors with default values
Rsim.state <- rsim.state(Rsim.params)
# Create stanza object
Rsim.stanzas <- rsim.stanzas(REco.params, Rsim.state, Rsim.params)

}
