% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim.R
\name{rsim.state}
\alias{rsim.state}
\title{Initialize starting state matrix for Rsim scenario}
\usage{
rsim.state(params)
}
\arguments{
\item{params}{Rsim parameter object generated by \code{rsim.params}.}
}
\value{
S3 object class \code{Rsim.state}, a list of 3 objects:
\itemize{
 \item{\code{Biomass}, numeric vector of initial biomass by biomass group}
 \item{\code{N}, numeric vector of  by biomass group}
 \item{\code{Ftime}, numeric vector of initial feeding time parameter by biomass group}
}
}
\description{
Creates a matrix of starting state variables used by Rsim, including initial 
biomass and numbers by group, as well as feeding time (Ftime). Default initial 
values: N = 0, Ftime = 1
}
\examples{
# Read in Rpath parameter file and generate model object
Rpath <- rpath(AB.params)
# Create dynamic parameters from Rpath model
Rsim.params <- rsim.params(Rpath)
# Create forcing matrix with default values
Rsim.state <- rsim.state(Rsim.params)
}
