% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosim_multistep.R
\name{rsim.step}
\alias{rsim.step}
\title{Initial function to step through ecosim}
\usage{
rsim.step(Rsim.scenario, Rsim.output, method = "AB", year.end)
}
\arguments{
\item{Rsim.scenario}{Rpath object created using rsim.scenario.}

\item{Rsim.output}{Rpath ecosim run created by the rsim.run() function.}

\item{method}{Which integration algorithim rsim.run will use (Currently only
supports 'AB')}

\item{step.end}{The end year for the current interval}
}
\value{
Returns an Rsim.output object.
}
\description{
Runs rsim.run at intervals and combines the run to one output
}
\seealso{
Other Rpath.functions: \code{\link{adjust.fishing}},
  \code{\link{adjust.forcing}},
  \code{\link{adjust.scenario}},
  \code{\link{check.rpath.params}},
  \code{\link{create.rpath.params}},
  \code{\link{extract.node}}, \code{\link{frate.table}},
  \code{\link{read.rpath.params}},
  \code{\link{rpath.stanzas}}, \code{\link{rpath}},
  \code{\link{rsim.params}}, \code{\link{rsim.plot}},
  \code{\link{rsim.scenario}}, \code{\link{stanzaplot}},
  \code{\link{webplot}}, \code{\link{write.Rpath}},
  \code{\link{write.Rsim}},
  \code{\link{write.rpath.params}}
}

