% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustments.R
\name{set.rsim.scene}
\alias{set.rsim.scene}
\title{Set Rsim.scenario parameters}
\usage{
set.rsim.scene(
  Rsim.scenario,
  params = NULL,
  start_state = NULL,
  forcing = NULL,
  fishing = NULL,
  stanzas = NULL
)
}
\arguments{
\item{Rsim.scenario}{Scenario object that contains all of the rsim rates and 
forcing functions generated by \code{rsim.scenario}.}

\item{params}{Rsim parameter object generated by \code{rsim.params}.}

\item{start_state}{Rsim starting values object generated by \code{rsim.state}}

\item{forcing}{Rsim forcing matrix object generated by \code{rsim.forcing}}

\item{fishing}{Rsim fishing matrix object generated by \code{rsim.fishing}}

\item{stanzas}{Rsim stanza parameters object generated by \code{rsim.stanzas}}
}
\value{
Returns an \code{Rsim.scenario} object with the new parameter.
}
\description{
Modifies the various parameters of the \code{rsim.scenario} object. Parameters 
that can be adjusted using this function are:
}
\concept{Set functions}
