% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecopath_plot.R
\name{stanzaplot}
\alias{stanzaplot}
\title{Plot routine for Ecopath multistanzas}
\usage{
stanzaplot(juvfile, StanzaNum, line.cols = c("black", "green", "blue", "red"))
}
\arguments{
\item{juvfile}{Rpath juvenile parameter file modified by the rpath.stanza() function.}

\item{StanzaNum}{The Stanza group's number in the stanzafile.}

\item{line.cols}{A vector of four colors used to represent the population biomass,
relative number, indvidual weights, and stanza separation lines.}
}
\value{
Creates a figure of the food web.
}
\description{
Plots the biomass composition of multistanza groups from an Rpath.stanzas object.
}
\seealso{
Other Rpath.functions: \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}};
  \code{\link{ecosim.plot}}; \code{\link{frate.adjust}};
  \code{\link{frate.table}}; \code{\link{rpath.stanzas}};
  \code{\link{rpath}}; \code{\link{rsim.params}};
  \code{\link{webplot}}; \code{\link{write.Rpath.sim}};
  \code{\link{write.Rpath}}
}

