% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopath_plot.R
\name{stanzaplot}
\alias{stanzaplot}
\title{Plot routine for Ecopath multistanzas}
\usage{
stanzaplot(
  Rpath.params,
  StanzaGroup,
  line.cols = c("black", "green", "blue", "red")
)
}
\arguments{
\item{Rpath.params}{R object containing the Rpath parameters.  This is generated
either by the create.rpath.params or read.rpath.params functions.}

\item{StanzaGroup}{The Stanza group's name to be plotted.}

\item{line.cols}{A vector of four colors used to represent the population biomass,
relative number, indvidual weights, and stanza separation lines.}
}
\value{
Creates a figure showing the break down of biomass and number per stanza.
}
\description{
Plots the biomass composition of multistanza groups from an Rpath.stanzas object.
}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{adjust.forcing}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{extract.node}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.stanzas}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.Rsim}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
