% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecopath_plot.R
\name{stanzaplot}
\alias{stanzaplot}
\title{Plot routine for Ecopath multistanzas}
\usage{
stanzaplot(Rpath.params, StanzaNum, line.cols = c("black", "green", "blue",
  "red"))
}
\arguments{
\item{Rpath.params}{R object containing the Rpath parameters modified by the
rpath.stanza function.}

\item{StanzaNum}{The Stanza group's number to be plotted.}

\item{line.cols}{A vector of four colors used to represent the population biomass,
relative number, indvidual weights, and stanza separation lines.}
}
\value{
Creates a figure showing the break down of biomass and number per stanza.
}
\description{
Plots the biomass composition of multistanza groups from an Rpath.stanzas object.
}
\seealso{
Other Rpath.functions: \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}};
  \code{\link{frate.adjust}}; \code{\link{frate.table}};
  \code{\link{read.rpath.params}};
  \code{\link{rpath.stanzas}}; \code{\link{rpath}};
  \code{\link{rsim.params}}; \code{\link{rsim.plot}};
  \code{\link{rsim.scenario}}; \code{\link{webplot}};
  \code{\link{write.Rpath.sim}}; \code{\link{write.Rpath}};
  \code{\link{write.rpath.params}}
}

