% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopath_plot.R
\name{stanzaplot}
\alias{stanzaplot}
\title{Plot routine for Ecopath multistanzas}
\usage{
stanzaplot(Rpath.params, StanzaGroup, line.cols = c("black", "green", "blue",
  "red"), ...)
}
\arguments{
\item{Rpath.params}{R object containing the Rpath parameters modified by the 
rpath.stanza function.}

\item{line.cols}{A vector of four colors used to represent the population biomass,
relative number, indvidual weights, and stanza separation lines.}

\item{StanzaNum}{The Stanza group's number to be plotted.}
}
\value{
Creates a figure showing the break down of biomass and number per stanza.
}
\description{
Plots the biomass composition of multistanza groups from an Rpath.stanzas object.
}
\seealso{
Other Rpath.functions: \code{\link{adjust.fishing}},
  \code{\link{adjust.forcing}},
  \code{\link{adjust.scenario}},
  \code{\link{check.rpath.params}},
  \code{\link{create.rpath.params}},
  \code{\link{extract.node}}, \code{\link{frate.table}},
  \code{\link{read.rpath.params}},
  \code{\link{rpath.stanzas}}, \code{\link{rpath}},
  \code{\link{rsim.params}}, \code{\link{rsim.plot}},
  \code{\link{rsim.scenario}}, \code{\link{rsim.step}},
  \code{\link{webplot}}, \code{\link{write.Rpath}},
  \code{\link{write.Rsim}},
  \code{\link{write.rpath.params}}
}

