% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ecopath.R
\name{webplot}
\alias{webplot}
\title{Plot routine for Ecopath food web}
\usage{
webplot(x, highlight = NULL, eco.name = attr(x, "eco.name"))
}
\arguments{
\item{highlight}{Box number to highlight connections.}

\item{eco.name}{Optional name of the ecosystem.  Default is the attribute of the
   rpath object.}

\item{Rpath.obj}{Rpath model created by the ecopath() function.}
}
\value{
Creates a figure of the food web.
}
\description{
Plots the food web associated with an Rpath object.
}
\seealso{
Other Rpath.functions: \code{\link{ecopath}};
  \code{\link{ecosim.init}}; \code{\link{ecosim.run}}
}

