% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopath_plot.R
\name{webplot}
\alias{webplot}
\title{Plot routine for Ecopath food web}
\usage{
webplot(Rpath.obj, eco.name = attr(Rpath.obj, "eco.name"),
  line.col = "grey", highlight = NULL, highlight.col = c("black", "red",
  "orange"), labels = FALSE, label.pos = NULL, label.num = FALSE,
  label.cex = 1, fleets = FALSE, type.col = "black", box.order = NULL,
  ...)
}
\arguments{
\item{Rpath.obj}{Rpath model created by the ecopath() function.}

\item{eco.name}{Optional name of the ecosystem.  Default is the eco.name attribute from the
rpath object.}

\item{line.col}{The color of the lines between nodes of the food web.}

\item{highlight}{Box number to highlight connections.}

\item{highlight.col}{Color of the connections to the highlighted group.}

\item{labels}{Logical whether or not to display group names.  If True and label.pos is Null, no 
points will be ploted, just label names.}

\item{label.pos}{A position specifier for the labels.  Values of 1, 2, 3, 4, respectively 
indicate positions below, to the left of, above, and to the right of the points. A null 
value will cause the labels to be ploted without the points (Assuming that labels = TRUE).}

\item{label.num}{Logical value indication whether group numbers should be used for labels 
instead of names.}

\item{label.cex}{The relative size of the labels within the plot.}

\item{fleets}{Logical value indicating whether or not to include fishing fleets in the food web.}

\item{type.col}{The color of the points cooresponding to the types of the group.  Can either be 
of length 1 or 4.  Color order will be living, primary producers, detrital, and fleet groups.}

\item{box.order}{Vector of box numbers to change the default plot order.  Must include all box numbers}

\item{highlight}{Set to the group number or name to highlight the connections of that group.}
}
\value{
Creates a figure of the food web.
}
\description{
Plots the food web associated with an Rpath object.
}
\seealso{
Other Rpath.functions: \code{\link{adjust.fishing}},
  \code{\link{adjust.forcing}},
  \code{\link{adjust.scenario}},
  \code{\link{check.rpath.params}},
  \code{\link{create.rpath.params}},
  \code{\link{extract.node}}, \code{\link{frate.table}},
  \code{\link{read.rpath.params}},
  \code{\link{rpath.stanzas}}, \code{\link{rpath}},
  \code{\link{rsim.params}}, \code{\link{rsim.plot}},
  \code{\link{rsim.scenario}}, \code{\link{rsim.step}},
  \code{\link{stanzaplot}}, \code{\link{write.Rpath}},
  \code{\link{write.Rsim}},
  \code{\link{write.rpath.params}}
}

