% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{write.rpath.params}
\alias{write.rpath.params}
\title{Write Rpath parameters to .csv files}
\usage{
write.rpath.params(Rpath.params, eco.name, path = "")
}
\arguments{
\item{Rpath.params}{R object containing the Rpath parameters.  Most likely this
was created using create.rpath.params or read.rpath.params.}

\item{eco.name}{ecosystem name that will be included in all the file names.}

\item{path}{location for the output files.}
}
\value{
Outputs a series of .csv files named by the provided eco.name and the 
 parameters they represent.  For example the model parameters will be named 
 "eco.name_model.csv".
}
\description{
Creates a series of .csv files from an Rpath.param object.
}
\seealso{
Other Rpath.functions: \code{\link{adjust.fishing}},
  \code{\link{adjust.scenario}},
  \code{\link{check.rpath.param}},
  \code{\link{create.rpath.param}},
  \code{\link{frate.table}},
  \code{\link{read.rpath.params}},
  \code{\link{rpath.stanzas}}, \code{\link{rpath}},
  \code{\link{rsim.params}}, \code{\link{rsim.plot}},
  \code{\link{rsim.scenario}}, \code{\link{stanzaplot}},
  \code{\link{webplot}}, \code{\link{write.Rpath.sim}},
  \code{\link{write.Rpath}}
}

