% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write.Rpath}
\alias{write.Rpath}
\title{Write function for Rpath food web object}
\usage{
write.Rpath(x, file = NA, morts = F)
}
\arguments{
\item{x}{Rpath model object generated by \code{rpath()}.}

\item{file}{file name for resultant file.  Need to specify ".csv" or ".RData".}

\item{morts}{Logical value whether to output basic parameters or mortalities.}
}
\value{
Writes a list object or file with the basic parameters or mortalities from an Rpath object.
}
\description{
Outputs basic parameters or mortalities from a \code{rpath()} object to a .csv or .RData file.
}
\examples{
# Read in Rpath parameter file, generate model object
x <- rpath(AB.params)
# Saves basic model parameters for all groups to the root directory of users project
write.Rpath(x, file = "output.csv", morts = F)
# Calculates natural mortality (M0), fishing mortality by fleet, and predation mortality by predator (M2) on each group
# Writes output to an Rdata file in the root directory of users project
write.Rpath(x, file = here::here("output.RData"), morts = T)


}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{adjust.forcing}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{extract.node}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.stanzas}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
