% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rpath_methods.R
\name{write.Rpath}
\alias{write.Rpath}
\title{Write function for Ecopath object}
\usage{
write.Rpath(x, file, morts = F, ...)
}
\arguments{
\item{x}{Rpath model created by the ecopath() function.}

\item{file}{file name for resultant .csv file.  Be sure to include ".csv".}

\item{morts}{Logical value whether to output basic parameters or mortalities.}
}
\value{
Writes a .csv file with the basic parameters or mortalities from an Rpath object.
}
\description{
Outputs basic parameters or mortalities to a .csv file.
}
\seealso{
Other Rpath.functions: \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}}; \code{\link{ecopath}};
  \code{\link{ecosim.init}}; \code{\link{ecosim.run}};
  \code{\link{webplot}}
}

