% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write.Rpath}
\alias{write.Rpath}
\title{Write function for Rpath food web object}
\usage{
write.Rpath(x, file = NA, morts = F)
}
\arguments{
\item{x}{Rpath model object generated by \code{rpath()}.}

\item{file}{file name for resultant file.  Need to specify ".csv" or ".RData".}

\item{morts}{Logical value whether to output basic parameters or mortalities.}
}
\value{
Writes a list object or file with the basic parameters or mortalities from an Rpath object.

If \code{morts} = F, output reflects the basic parameters of the mass-balanced model, namely:
\item{Group}{Vector of group names}
\item{type}{Vector of group types: 0 = consumer, 1 = producer, 2 = detritus, 3 = fishery}
\item{TL}{Vector of trophic level for each group, calculated based on mass balance}
\item{Biomass}{Vector of biomass values for each modeled group}
\item{PB}{Vector of production:biomass ratios for each modeled group}
\item{QB}{Vector of consumption:biomass ratios for each modeled group}
\item{EE}{Vector of ecotrophic efficiency values for each modeled group}
\item{GE}{Vector of gross efficiency values (production:comsumption ratios) for each modeled group}
\item{Removals}{Vector of total fishery removals for each modeled group}

If \code{morts} = T, output reflects calculated mortality values for each group. 
The number of columns will vary based on how many predators are in the model.
\item{Group}{Vector of group names}
\item{M0}{Vector of natural mortality for each group. Natural mortality = PB * (1-EE)}
\item{Fmort}{Vector of fishing mortality for each group}
\item{M2.XX}{Vector of predation mortality, where XX is the name of a predator group}
}
\description{
Outputs basic parameters or mortalities from a \code{rpath()} object to a .csv or .RData file.
}
\examples{
# Read in Rpath parameter file, generate model object
x <- rpath(AB.params)
# Saves basic model parameters for all groups to the root directory of users project
write.Rpath(x, file = "output.csv", morts = F)
# Calculates natural mortality (M0), fishing mortality by fleet, and predation mortality by predator (M2) on each group
# Writes output to an Rdata file in the root directory of users project
write.Rpath(x, file = here::here("output.RData"), morts = T)


}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{adjust.forcing}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{extract.node}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.stanzas}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
