% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write.Rpath}
\alias{write.Rpath}
\title{Write function for Ecopath object}
\usage{
write.Rpath(x, file, morts = F, ...)
}
\arguments{
\item{x}{Rpath model created by the ecopath() function.}

\item{file}{file name for resultant .csv file.  Be sure to include ".csv".}

\item{morts}{Logical value whether to output basic parameters or mortalities.}
}
\value{
Writes a .csv file with the basic parameters or mortalities from an Rpath object.
}
\description{
Outputs basic parameters or mortalities to a .csv file.
}
\seealso{
Other Rpath.functions: \code{\link{adjust.fishing}},
  \code{\link{adjust.forcing}},
  \code{\link{adjust.scenario}},
  \code{\link{check.rpath.params}},
  \code{\link{create.rpath.params}},
  \code{\link{extract.node}}, \code{\link{frate.table}},
  \code{\link{read.rpath.params}},
  \code{\link{rpath.stanzas}}, \code{\link{rpath}},
  \code{\link{rsim.params}}, \code{\link{rsim.plot}},
  \code{\link{rsim.scenario}}, \code{\link{rsim.step}},
  \code{\link{stanzaplot}}, \code{\link{webplot}},
  \code{\link{write.Rsim}},
  \code{\link{write.rpath.params}}
}

