% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write.Rpath.sim}
\alias{write.Rpath.sim}
\title{Write function for Ecosim object}
\usage{
write.Rpath.sim(x, file, ...)
}
\arguments{
\item{x}{Rpath.sim object created by the ecosim.run() function.}

\item{file}{file name for resultant .csv file.  Be sure to include ".csv".}
}
\value{
Writes a .csv file with the start and end biomass and catch per group
from an Rpath.sim object.
}
\description{
Outputs start/end biomass and catch to a .csv file.
}
\seealso{
Other Rpath.functions: \code{\link{Rsim.params}};
  \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}}; \code{\link{ecopath}};
  \code{\link{ecosim.plot}}; \code{\link{frate.adjust}};
  \code{\link{frate.table}}; \code{\link{rpath.stanzas}};
  \code{\link{stanzaplot}}; \code{\link{webplot}};
  \code{\link{write.Rpath}}
}

