% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write.Rsim}
\alias{write.Rsim}
\title{Write function for Ecosim object}
\usage{
write.Rsim(Rsim.output, file, ...)
}
\arguments{
\item{Rsim.output}{object created by the rsim.run() function.}

\item{file}{file name for resultant .csv file.  Be sure to include ".csv".}
}
\value{
Writes a .csv file with the start and end biomass and catch per group
from an Rpath.sim object.
}
\description{
Outputs start/end biomass and catch to a .csv file.
}
\seealso{
Other Rpath.functions: \code{\link{adjust.fishing}};
  \code{\link{adjust.scenario}};
  \code{\link{check.rpath.param}};
  \code{\link{create.rpath.param}};
  \code{\link{frate.table}};
  \code{\link{read.rpath.param}};
  \code{\link{rpath.stanzas}}; \code{\link{rpath}};
  \code{\link{rsim.params}}; \code{\link{rsim.plot}};
  \code{\link{rsim.scenario}}; \code{\link{stanzaplot}};
  \code{\link{webplot}}; \code{\link{write.Rpath}};
  \code{\link{write.rpath.param}}
}

