% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write.Rsim}
\alias{write.Rsim}
\title{Write function for Ecosim object}
\usage{
write.Rsim(Rsim.output, file = NA)
}
\arguments{
\item{Rsim.output}{object created by the rsim.run() function.}

\item{file}{file name for resultant .csv file.  Be sure to include ".csv".}
}
\value{
Writes a .csv file with the start and end biomass and catch per group
from an Rpath.sim object.
}
\description{
Outputs start/end biomass and catch to a .csv file.
}
\seealso{
Other Rpath functions: 
\code{\link{MTI}()},
\code{\link{adjust.forcing}()},
\code{\link{check.rpath.params}()},
\code{\link{create.rpath.params}()},
\code{\link{extract.node}()},
\code{\link{frate.table}()},
\code{\link{read.rpath.params}()},
\code{\link{rpath}()},
\code{\link{rpath.groups}()},
\code{\link{rpath.stanzas}()},
\code{\link{stanzaplot}()},
\code{\link{webplot}()},
\code{\link{write.Rpath}()},
\code{\link{write.rpath.params}()}
}
\concept{Rpath functions}
